refactor(runtime,backends): invert driver/backend relationship
Move VMM mechanics from the legacy direct VmBackend shells into the
VmmDriver implementations. Delete FirecrackerBackend, HvfBackend,
LibkrunBackend, and QemuBackend; keep helper machinery in driver-local
legacy modules. Every selectable microVM backend now reaches production
through the blanket impl VmBackend for WorkloadRunner<D: VmmDriver, ...>.

WasmBackend remains a documented direct-VmBackend exemption (WASI module,
not a microVM). QEMU stays an opt-in Tier-2 dev/test backend, never
workload-bearing.

Verification:
- cargo nextest run --workspace: 11275 passed, 26 skipped
- cargo clippy --workspace --all-targets -- -D warnings: clean
- cargo xtask check-claim-catalog: clean
- Migration boundary recorded in MIGRATION-269.md
