+++ crates/mvm-backends/src/driver/mod.rs
@@ -2,9 +2,14 @@
 
 pub mod fc;
 pub mod hvf;
+pub mod hvf_bootargs;
 pub mod hvf_restore;
 pub mod libkrun;
 pub mod qemu;
+
+pub mod hvf_legacy;
+pub mod libkrun_legacy;
+pub mod qemu_legacy;
 
 pub use fc::FcDriver;
 pub use hvf::HvfDriver;
