--- crates/mvm-backends/src/legacy/libkrun.rs
@@ -1,2505 +1,0 @@
-//! libkrun backend for mvm.
-//!
-//! Tier 2 microVM backend that runs on
-//! Linux KVM and macOS Apple Silicon. The lifecycle
-//! delegates to a per-VM `mvm-libkrun-supervisor` subprocess
-//! rather than calling libkrun in-process: `krun_start_enter` calls
-//! `exit()` on the host process when the guest powers off, so any
-//! in-process registry would tear down sibling guests. One process per
-//! VM scopes the `exit()` to that supervisor and lets the parent
-//! `mvmctl` survive a guest shutdown.
-//!
-//! ## Lifecycle
-//!
-//! - `start` writes `~/.mvm/vms/<name>/{rootfs.ref,kernel.ref}` runtime
-//!   metadata (so `mvmctl console` can find the artifacts), serializes a
-//!   [`SupervisorConfig`], spawns `mvm-libkrun-supervisor` with the JSON
-//!   on stdin, and waits up to `PID_FILE_TIMEOUT` for the supervisor
-//!   to write its PID file. Returns once the supervisor is running or
-//!   exits with an error if the spawn fails or PID file never appears.
-//! - `stop` reads `<vm_state_dir>/libkrun.pid`, arms a host process-exit
-//!   observer before `SIGTERM`, and falls back to bounded polling plus
-//!   `SIGKILL` if the process does not exit within `STOP_TIMEOUT`.
-//! - `status` reads the PID file and probes with `kill(pid, 0)`.
-//! - `list` walks `~/.mvm/vms/*/libkrun.pid`.
-
-use anyhow::{Context, Result, anyhow, bail};
-use libkrun_sys::{
-    BridgeRestartPolicy, KrunContext, SupervisorAttachConfig, SupervisorBaseConfig,
-    SupervisorConfig,
-};
-use mvm_core::config::{vm_console_log, vm_libkrun_pid, vm_state_dir, vms_dir};
-use mvm_core::kernel_format::KernelFormat;
-use mvm_core::vm_backend::{
-    BackendKind, BackendSecurityProfile, ClaimStatus, GuestChannelInfo, LayerCoverage,
-    ResourceControls, SnapshotCapability, StandbyClaim, StandbyError, StandbyHandle, StandbySpec,
-    StandbyState, StartMode, VmBackend, VmCapabilities, VmId, VmInfo, VmStartConfig, VmStatus,
-    WarmStartError, WarmStartOutcome,
-};
-use mvm_vmm::host::snapshot_upper::SnapshotUpper;
-use mvm_vmm::host::ui;
-use std::io::Write;
-use std::os::unix::net::UnixStream;
-use std::path::{Path, PathBuf};
-use std::process::{Command, Stdio};
-use std::time::{Duration, Instant};
-
-/// libkrun backend (Linux KVM / macOS Hypervisor.framework).
-pub struct LibkrunBackend;
-
-use mvm_vmm::host::substitution_spawn::EndpointGuard;
-
-/// Spawn the per-VM egress endpoint for libkrun workloads. Secret-bound runs
-/// get the WireRequest substitution path; secret-free runs can opt into the
-/// raw vsock relay when the resolved policy allows egress. Deny-all,
-/// secret-free runs keep the endpoint defused because the guest will not be
-/// told to start the egress client.
-fn spawn_libkrun_egress_endpoint_if_needed(
-    vm_name: &str,
-    state_dir: &Path,
-    config_tenant: &str,
-    network_policy: &mvm_core::network_policy::NetworkPolicy,
-) -> Result<EndpointGuard> {
-    use mvm_vmm::host::substitution_spawn::{
-        EndpointTransport, SubstitutionSpawnParams, spawn_substitution_endpoint,
-    };
-    let default_redaction = mvm_core::policy::RedactionPolicy::default();
-    let decoded = mvm_vmm::host::egress_shared::decode_plan_secrets_from_state(state_dir)?;
-    let (secrets, redaction, tenant): (&[_], &mvm_core::policy::RedactionPolicy, &str) =
-        match &decoded {
-            Some((secrets, redaction, tenant)) => (secrets.as_slice(), redaction, tenant.as_str()),
-            None => (
-                &[],
-                &default_redaction,
-                if config_tenant.is_empty() {
-                    "local"
-                } else {
-                    config_tenant
-                },
-            ),
-        };
-    if secrets.is_empty() && !network_policy.allows_egress() {
-        return Ok(EndpointGuard::defused());
-    }
-    spawn_substitution_endpoint(SubstitutionSpawnParams {
-        vm_name,
-        state_dir,
-        tenant,
-        secrets,
-        redaction,
-        transport: EndpointTransport::Uds {
-            path: mvm_core::config::vm_vsock_port_socket(vm_name, mvm_agentd::vsock::EGRESS_PORT),
-        },
-        terminator_listen: None,
-        tls_intermediate: None,
-        network_policy: Some(network_policy),
-        raw_egress: secrets.is_empty(),
-        resolver_remote: None,
-        binding_store_dir: None,
-    })?;
-    Ok(EndpointGuard::new(vm_name))
-}
-
-/// Kernel cmdline token that turns on the in-guest vsock egress client.
-/// Emitted only when outbound egress is allowed and the workload carries no
-/// bound secrets, so the raw SOCKS path never contends with the
-/// substitution endpoint for the shared egress port.
-fn vsock_egress_cmdline_token(config: &VmStartConfig, state_dir: &Path) -> Option<String> {
-    (config.network_policy.allows_egress()
-        && !mvm_vmm::host::egress_shared::state_has_bound_secrets(state_dir).unwrap_or(false))
-    .then(|| "mvm.vsock_egress=1".to_string())
-}
-
-/// How long [`LibkrunBackend::start`] waits for the supervisor to
-/// write its PID file before giving up and killing the child. Shared with
-/// the libkrun driver, which spawns the same supervisor binary.
-pub(crate) const PID_FILE_TIMEOUT: Duration = Duration::from_secs(5);
-/// After the PID file appears the supervisor still has to bind the
-/// per-port vsock listener socket; callers (the console attach,
-/// `shell_exec` via `connect_with_auto_start`) race that window and
-/// wrongly see the VM as "not running". `start` waits for the
-/// agent socket before returning so "ready" actually means reachable. Shared
-/// with the libkrun driver's boot path.
-pub(crate) const VSOCK_SOCKET_TIMEOUT: Duration = Duration::from_secs(10);
-
-/// How long [`LibkrunBackend::stop`] waits after `SIGTERM` before
-/// escalating to `SIGKILL`. Tight because libkrun's signal handling
-/// under `krun_start_enter` is empirically unreliable — when the
-/// supervisor is spawned via `std::process::Command` (the production
-/// `mvmctl up` path), SIGTERM often doesn't reach the in-supervisor
-/// `sigaction` handler before we escalate. The in-supervisor handler
-/// (see `libkrun_sys::install_shutdown_handler`) still helps the
-/// shell-stop case where SIGTERM is delivered cleanly (~200 ms); in
-/// the always-escalate cargo-spawn / launchd path a shorter timeout
-/// means `mvmctl stop` returns in 2 s instead of 5 s. Shared with the libkrun
-/// driver's `kill` escalation.
-pub const STOP_TIMEOUT: Duration = Duration::from_secs(2);
-const FORCE_KILL_TIMEOUT: Duration = Duration::from_millis(500);
-
-/// Open the per-VM guest-console capture sink. OUTPUT-ONLY by
-/// construction (write-only, create+truncate): the guest console
-/// streams here and NO host-readable fd is ever attached as console
-/// input. The sole interactive path into a guest is the interactive-gated
-/// Default kernel cmdline for libkrun-launched guests.
-/// `console=hvc0` matches libkrun's virtio-console wiring;
-/// `root=/dev/vda rw init=/init` matches Apple Container's
-/// boot for the same Nix-built rootfs layout.
-pub(crate) const DEFAULT_CMDLINE: &str = "console=hvc0 root=/dev/vda rw init=/init";
-pub(crate) const VERITY_CMDLINE: &str = "console=hvc0";
-
-/// Build the supervisor config for one VM, lifting
-/// the audit-substrate resolution into the shared `audit_substrate`
-/// module so any adopting backend shares one source of truth (and the
-/// future `NetworkProvider` trait extraction is mechanical).
-///
-/// **Do not log** `config.plan_json` or `config.bundle_json` — they
-/// may carry secret bindings, env vars, or policy refs that resolve
-/// to credentials. They're opaque transport bytes; the supervisor
-/// re-verifies the signed envelope before trusting any decoded field.
-/// Workload-**independent** KrunContext: kernel + resources + cmdline + vsock wiring +
-/// the configured direct-vsock transport. **No rootfs** (the cold path sets the
-/// workload rootfs; a prelaunched standby leaves it `None` until claim) and **no
-/// user volumes**.
-/// Shared verbatim by `build_supervisor_config` (cold) and `standby_base_config` (warm)
-/// so the two launch paths can't drift.
-fn krun_context_base(
-    name: &str,
-    kernel: &str,
-    vcpus: u8,
-    mem_mib: u32,
-    cmdline: &str,
-    state_dir: &Path,
-) -> KrunContext {
-    // KrunContext::new requires a rootfs arg; we null it immediately — the caller owns
-    // the rootfs decision (Some for cold boot, None for a standby).
-    let mut krun = KrunContext::new(name, kernel, "")
-        .with_resources(vcpus, mem_mib)
-        .with_cmdline(cmdline)
-        .with_vsock_socket_dir(state_dir.to_string_lossy().into_owned())
-        .add_vsock_port(mvm_agentd::vsock::GUEST_AGENT_PORT)
-        // Workload exit-code capture (listen=false → host binds).
-        .add_host_listen_port(mvm_agentd::vsock::WORKLOAD_EXIT_PORT)
-        // Egress substitution channel (listen=false → host binds). Registered
-        // unconditionally; fail-closed: when the plan carries no secrets nothing
-        // binds the UDS, so a stray guest dial gets ECONNREFUSED.
-        .add_host_listen_port(mvm_agentd::vsock::EGRESS_PORT)
-        // Host-services broker channel (listen=false → host binds). Registered
-        // unconditionally so libkrun proxies the guest's connect; fail-closed:
-        // nothing binds the UDS until the per-VM broker subprocess is spawned,
-        // so a stray guest dial gets ECONNREFUSED.
-        .add_host_listen_port(mvm_agentd::vsock::BROKER_PORT);
-    krun.rootfs_path = None;
-    // The active libkrun transport is direct-vsock only; keep the workload
-    // base aligned with the builder path so no guest-NIC helper can be
-    // reopened through a different launch site.
-    krun.with_vsock_direct()
-}
-
-/// Translate a backend-agnostic [`StandbySpec`] into the `SupervisorBaseConfig`
-/// (workload-independent) the prelaunched supervisor reads on stdin. `rootfs_path` stays
-/// `None` — the workload rootfs arrives in the attach at claim.
-fn standby_base_config(spec: &StandbySpec) -> SupervisorBaseConfig {
-    let krun = krun_context_base(
-        &spec.id,
-        &spec.kernel_path,
-        spec.vcpus,
-        spec.mem_mib,
-        DEFAULT_CMDLINE,
-        Path::new(&spec.vm_state_dir),
-    );
-    SupervisorBaseConfig {
-        krun,
-        vm_state_dir: spec.vm_state_dir.clone(),
-        pid_file_name: None,
-        signing_key_path: spec.signing_key_path.clone().into(),
-        signer_id: spec.signer_id.clone(),
-        binding_nonce: spec.binding_nonce.clone(),
-        control_socket_path: spec.control_socket.clone().into(),
-        bridge_restart_policy: BridgeRestartPolicy::HardFail,
-    }
-}
-
-/// Translate a [`StandbyClaim`] into the `SupervisorAttachConfig`, echoing the
-/// standby's binding nonce. `plan_json`/`bundle_json` decode to `serde_json::Value`
-/// carriers (the same shape `SupervisorConfig.plan` uses).
-fn standby_attach_config(
-    claim: &StandbyClaim,
-    binding_nonce: String,
-) -> std::result::Result<SupervisorAttachConfig, StandbyError> {
-    let plan: serde_json::Value = serde_json::from_str(&claim.plan_json)
-        .map_err(|e| StandbyError::ClaimFailed(format!("decode plan_json: {e}")))?;
-    let bundle = match &claim.bundle_json {
-        Some(b) => Some(
-            serde_json::from_str(b)
-                .map_err(|e| StandbyError::ClaimFailed(format!("decode bundle_json: {e}")))?,
-        ),
-        None => None,
-    };
-    let network_policy = Some(
-        serde_json::to_value(&claim.network_policy)
-            .map_err(|e| StandbyError::ClaimFailed(format!("encode network_policy: {e}")))?,
-    );
-    Ok(SupervisorAttachConfig {
-        binding_nonce,
-        rootfs_path: claim.rootfs_path.clone(),
-        tenant_id: claim.tenant_id.clone(),
-        audit_dir: claim.audit_dir.clone(),
-        gateway_audit_socket: claim.gateway_audit_socket.clone(),
-        gateway_events_socket: claim.gateway_events_socket.clone(),
-        plan,
-        bundle,
-        network_policy,
-        transparent_terminator_port: claim.start_config.as_ref().map(|config| {
-            mvm_vmm::host::egress_redirect::terminator_port_for_vm_name(&config.name)
-        }),
-    })
-}
-
-pub(crate) fn libkrun_kernel_for_host(kernel: &str) -> Result<(String, KernelFormat)> {
-    let path = Path::new(kernel);
-    if cfg!(target_arch = "x86_64") && path.exists() {
-        let loadable = mvm_vmm::host::fc_kernel::ensure_fc_loadable_kernel(path)
-            .with_context(|| format!("preparing x86_64 libkrun-loadable kernel from {kernel}"))?;
-        return Ok((loadable.to_string_lossy().into_owned(), KernelFormat::Elf));
-    }
-    Ok((kernel.to_string(), KernelFormat::Raw))
-}
-
-fn libkrun_verity_initrd_path(config: &VmStartConfig) -> Option<PathBuf> {
-    config
-        .verity_path
-        .as_deref()
-        .zip(config.roothash.as_deref())
-        .and_then(|_| {
-            Path::new(&config.rootfs_path)
-                .parent()
-                .map(|p| p.join("rootfs.initrd"))
-        })
-        .filter(|p| p.exists())
-}
-
-fn libkrun_effective_initrd(config: &VmStartConfig) -> Option<PathBuf> {
-    config
-        .initrd_path
-        .as_ref()
-        .map(PathBuf::from)
-        .or_else(|| libkrun_verity_initrd_path(config))
-}
-
-fn libkrun_verity_enabled(config: &VmStartConfig) -> bool {
-    config.verity_path.is_some()
-        && config.roothash.is_some()
-        && libkrun_effective_initrd(config).is_some()
-}
-
-fn libkrun_runtime_overlay(config: &VmStartConfig) -> Option<(&str, &str, &str)> {
-    Some((
-        config.runtime_overlay_path.as_deref()?,
-        config.runtime_overlay_verity_path.as_deref()?,
-        config.runtime_overlay_roothash.as_deref()?,
-    ))
-}
-
-fn libkrun_verity_cmdline_args(config: &VmStartConfig) -> Option<String> {
-    let rootfs_hash = config.roothash.as_deref()?;
-    let base = format!("mvm.roothash={rootfs_hash} mvm.data=/dev/vdb mvm.hash=/dev/vdc");
-    match libkrun_runtime_overlay(config) {
-        Some((_, _, overlay_hash)) => Some(format!(
-            "{base} mvm.runtime_roothash={overlay_hash} mvm.runtime_data=/dev/vdd mvm.runtime_hash=/dev/vde"
-        )),
-        None => Some(base),
-    }
-}
-
-fn ensure_libkrun_runtime_source_supported(config: &VmStartConfig) -> Result<()> {
-    if config.runtime_source_policy != mvm_core::vm_backend::RuntimeSourcePolicy::RequiredOverlay {
-        return Ok(());
-    }
-    // A sealed boot (verity metadata present) must be fully verity capable — a
-    // missing initrd fails closed rather than downgrading to an unverified root —
-    // and carries the dm-verity overlay triple its initramfs mounts. A non-verity
-    // dev boot instead mounts a plain read-only overlay from `/dev/vdb`.
-    let verity_intended = config.roothash.is_some() || config.verity_path.is_some();
-    if verity_intended {
-        if !libkrun_verity_enabled(config) {
-            bail!(
-                "required-overlay libkrun boot requires verity metadata plus an initrd \
-                 (`--initrd` or sibling rootfs.initrd)"
-            );
-        }
-        if libkrun_runtime_overlay(config).is_none() {
-            bail!("required-overlay libkrun boot requires the runtime overlay artifact triple");
-        }
-    } else if mvm_vmm::host::boot_config::non_verity_overlay_ext4(config).is_none() {
-        bail!(
-            "required-overlay libkrun boot requires the runtime overlay artifact triple \
-             (a non-verity boot mounts it as a plain read-only /dev/vdb)"
-        );
-    }
-    Ok(())
-}
-
-/// Assemble the guest kernel cmdline: the verity or default base string plus
-/// every optional token layered on top, in the exact order the guest
-/// `/init` expects them.
-fn build_guest_cmdline(config: &VmStartConfig, state_dir: &Path) -> String {
-    // Append the `mvm.uvols=` param so the dev VM's `mvm-host-vm-init`
-    // mounts user volumes at their guest paths (no-op when there are
-    // none; harmless for workload guests whose `/init` ignores it).
-    let mut cmdline = if libkrun_verity_enabled(config) {
-        VERITY_CMDLINE.to_string()
-    } else {
-        DEFAULT_CMDLINE.to_string()
-    };
-    if let Some(uvols) = mvm_core::vm_backend::encode_user_volumes_cmdline(&config.volumes) {
-        cmdline.push(' ');
-        cmdline.push_str(&uvols);
-    }
-    if let Some(token) = mvm_vmm::host::egress_bridge::verb_grant_cmdline_token(&config.name) {
-        cmdline.push(' ');
-        cmdline.push_str(&token);
-    }
-    if let Some(token) = mvm_vmm::host::egress_bridge::require_grant_cmdline_token(&config.name) {
-        cmdline.push(' ');
-        cmdline.push_str(&token);
-    }
-    if let Some(verity_args) = libkrun_verity_enabled(config)
-        .then(|| libkrun_verity_cmdline_args(config))
-        .flatten()
-    {
-        cmdline.push(' ');
-        cmdline.push_str(&verity_args);
-    }
-    // Non-verity boots carry the runtime overlay as a plain read-only
-    // `/dev/vdb`; emit the token its `/init` mounts from. Verity boots already
-    // emitted the dm-verity variant above.
-    if !libkrun_verity_enabled(config)
-        && let Some(overlay_args) = mvm_vmm::host::boot_config::build_runtime_overlay_cmdline_args(
-            None,
-            mvm_vmm::host::boot_config::non_verity_overlay_ext4(config).is_some(),
-        )
-    {
-        cmdline.push(' ');
-        cmdline.push_str(&overlay_args);
-    }
-    cmdline.push(' ');
-    cmdline.push_str(&mvm_core::vm_backend::encode_runtime_source_policy_cmdline(
-        config.runtime_source_policy,
-    ));
-    // Vsock-only guests have no config drive, so the grant trust anchor rides
-    // the cmdline instead.
-    if let Some(token) = mvm_vmm::host::egress_bridge::host_signer_pub_cmdline_token(&config.name) {
-        cmdline.push(' ');
-        cmdline.push_str(&token);
-    }
-    if let Some(token) = vsock_egress_cmdline_token(config, state_dir) {
-        cmdline.push(' ');
-        cmdline.push_str(&token);
-    }
-    // The L3 tunnel and the socket-aware egress client are alternatives, not
-    // layers: the mode the plan admitted decides which one the guest starts.
-    if let Some(token) = mvm_vmm::host::egress_shared::l3_cmdline_token(config) {
-        cmdline.push(' ');
-        cmdline.push_str(&token);
-    }
-    cmdline
-}
-
-/// Attach the workload rootfs (plus its dm-verity sidecar and the runtime
-/// overlay disk) to `krun`. Verity boots carry `root`/`root-verity` disks
-/// and an initramfs; non-verity boots set `rootfs_path` directly and, when a
-/// runtime overlay is present, attach it as a plain read-only `/dev/vdb`.
-fn attach_workload_rootfs(mut krun: KrunContext, config: &VmStartConfig) -> KrunContext {
-    if libkrun_verity_enabled(config) {
-        krun.initramfs_path =
-            libkrun_effective_initrd(config).map(|p| p.to_string_lossy().into_owned());
-        krun = krun
-            .add_disk("root", config.rootfs_path.clone(), true)
-            .add_disk(
-                "root-verity",
-                config
-                    .verity_path
-                    .clone()
-                    .expect("verity-enabled libkrun boot must carry a verity sidecar"),
-                true,
-            );
-        if let Some((overlay_path, overlay_verity_path, _)) = libkrun_runtime_overlay(config) {
-            krun = krun.add_disk("runtime", overlay_path, true).add_disk(
-                "runtime-verity",
-                overlay_verity_path,
-                true,
-            );
-        }
-    } else {
-        krun.rootfs_path = Some(config.rootfs_path.clone());
-        // A non-verity dev boot has no initramfs to mount the runtime overlay,
-        // so attach it as a plain read-only virtio-blk device. `rootfs_path`
-        // takes /dev/vda, so this first extra disk is /dev/vdb — the device the
-        // `mvm.runtime_data=` token names for the guest `/init`.
-        if let Some(overlay) = mvm_vmm::host::boot_config::non_verity_overlay_ext4(config) {
-            krun = krun.add_disk("runtime", overlay.to_string(), true);
-        }
-    }
-    krun
-}
-
-/// Attach the user-supplied volumes (`--volume` / `MVM_VOLUMES`) to `krun`.
-/// A directory share is a virtio-fs device; a disk image is an extra
-/// virtio-blk device. Tag/id `uvol{idx}` is the coordination key the guest
-/// mount manifest uses to mount each at its requested guest path. Disk
-/// images are sparse-created by the CLI orchestrator before start; a RO
-/// disk image is RO at the hypervisor (krun_add_disk takes read_only).
-///
-/// libkrun's `krun_add_virtiofs` has NO host-side read-only
-/// toggle, so a "read-only" virtio-fs share would only be ro by the
-/// guest's own mount flag — a compromised guest could remount it rw.
-/// Rather than make a false ro promise, refuse it and point at the
-/// hypervisor-enforced alternatives. (HVF/Firecracker enforce ro shares
-/// natively; this restriction is libkrun-specific.)
-fn attach_user_volumes(mut krun: KrunContext, config: &VmStartConfig) -> Result<KrunContext> {
-    for (idx, vol) in config.volumes.iter().enumerate() {
-        let tag = format!("uvol{idx}");
-        krun = match vol.kind {
-            mvm_core::vm_backend::VmVolumeKind::DirShare => {
-                if vol.read_only {
-                    bail!(
-                        "libkrun cannot enforce a read-only virtio-fs share ('{}' -> '{}'): \
-                         krun_add_virtiofs has no host-side read-only toggle, so a compromised \
-                         guest could remount it read-write. Use a read-only disk image \
-                         (host:/guest:SIZE) for hypervisor-enforced read-only, share it ':rw' if \
-                         writes are intended, or run on the HVF backend.",
-                        vol.host,
-                        vol.guest
-                    );
-                }
-                krun.add_virtio_fs(tag, vol.host.clone())
-            }
-            mvm_core::vm_backend::VmVolumeKind::Disk => {
-                krun.add_disk(tag, vol.host.clone(), vol.read_only)
-            }
-        };
-    }
-    Ok(krun)
-}
-
-/// Assemble the per-workload [`KrunContext`]: the workload-independent base
-/// (kernel + resources + vsock + endpoint), shared verbatim with the standby
-/// spawn so the two paths can't drift, plus the cold-path additions —
-/// console capture, the workload rootfs, the
-/// dev-console data ports, and any user-supplied volumes.
-fn build_workload_krun_context(
-    config: &VmStartConfig,
-    kernel: &str,
-    kernel_format: KernelFormat,
-    vcpus: u8,
-    cmdline: &str,
-    state_dir: &Path,
-) -> Result<KrunContext> {
-    let mut krun = krun_context_base(
-        &config.name,
-        kernel,
-        vcpus,
-        config.memory_mib,
-        cmdline,
-        state_dir,
-    );
-    krun = krun
-        .with_kernel_format(kernel_format)
-        .with_console_output(vm_console_log(&config.name).display().to_string());
-    krun = attach_workload_rootfs(krun, config);
-
-    // A dev-accessible managed machine (`machine run -t` / `machine shell` /
-    // `up --console`) pre-opens the interactive-console data range. libkrun
-    // binds one host UDS listener per registered port (`listen=true`, host
-    // dials → guest's vsock server), so the agent's dynamic
-    // `CONSOLE_PORT_BASE + session_id` data port is unreachable unless it was
-    // declared here. Left out of `krun_context_base` so the warm-standby base
-    // (no `VmStartConfig` yet) stays unchanged.
-    if config.dev_console {
-        for port in mvm_agentd::vsock::dev_console_data_ports() {
-            krun = krun.add_vsock_port(port);
-        }
-    }
-
-    attach_user_volumes(krun, config)
-}
-
-/// Validate the tenant label and resolve the gateway-bridge audit substrate
-/// (paths).
-///
-/// `tenant_id` flows into the per-VM audit-chain path
-/// (`workload_audit_path`) that the host-services broker spawns against —
-/// even on the legacy/no-bridge path where the substrate stays all-None. So
-/// the tenant is validated here regardless of `plan_json`: an unsafe tenant
-/// (e.g. a path-traversal label) must be refused before `start()` reaches
-/// the broker spawn, independent of whether the gateway bridge is engaged. A
-/// path-injection guard, not just the bridge's input check.
-///
-/// Substrate resolution is gated on `plan_json` — the bridge supervisor's
-/// actual input — NOT on `tenant_id`: a substrate without a plan is what the
-/// supervisor rejects as "cfg.plan missing on bridge path". An admitted
-/// workload carries both `tenant_id` and `plan_json`, so the default
-/// workload path reaches the enforcing bridge; dev/builder starts keep
-/// `plan_json` absent and stay on the legacy `run_supervisor` path with an
-/// all-None substrate. (`compute_audit_substrate` re-validates the tenant —
-/// harmless after the guard above.)
-fn resolve_audit_substrate(
-    config: &VmStartConfig,
-) -> Result<mvm_vmm::host::audit_substrate::AuditSubstrate> {
-    if let Some(tenant) = config.tenant_id.as_deref() {
-        mvm_vmm::host::audit_substrate::validate_tenant_id(tenant)?;
-    }
-    Ok(if config.plan_json.is_some() {
-        mvm_vmm::host::audit_substrate::compute_audit_substrate(
-            &config.name,
-            config.tenant_id.as_deref(),
-        )?
-    } else {
-        mvm_vmm::host::audit_substrate::AuditSubstrate::default()
-    })
-}
-
-/// Parse the signed-plan + bundle envelopes from `VmStartConfig`.
-/// `libkrun_sys::SupervisorConfig` carries them as `Option<serde_json::Value>`
-/// so the supervisor can re-verify the envelope without typed coupling
-/// to `mvm_core::plan` at the parse boundary.
-fn parse_plan_and_bundle(
-    config: &VmStartConfig,
-) -> Result<(Option<serde_json::Value>, Option<serde_json::Value>)> {
-    let plan = match config.plan_json.as_deref() {
-        Some(s) => Some(
-            serde_json::from_str(s)
-                .map_err(|e| anyhow!("parse VmStartConfig.plan_json as JSON: {e}"))?,
-        ),
-        None => None,
-    };
-    let bundle = match config.bundle_json.as_deref() {
-        Some(s) => Some(
-            serde_json::from_str(s)
-                .map_err(|e| anyhow!("parse VmStartConfig.bundle_json as JSON: {e}"))?,
-        ),
-        None => None,
-    };
-    Ok((plan, bundle))
-}
-
-fn build_supervisor_config(config: &VmStartConfig, state_dir: &Path) -> Result<SupervisorConfig> {
-    let kernel = config
-        .kernel_path
-        .as_deref()
-        .ok_or_else(|| anyhow!("libkrun backend requires a kernel path"))?;
-    let (kernel, kernel_format) = libkrun_kernel_for_host(kernel)?;
-    let vcpus = u8::try_from(config.cpus.clamp(1, u32::from(u8::MAX))).unwrap_or(u8::MAX);
-    let cmdline = build_guest_cmdline(config, state_dir);
-    let krun =
-        build_workload_krun_context(config, &kernel, kernel_format, vcpus, &cmdline, state_dir)?;
-    let substrate = resolve_audit_substrate(config)?;
-    let (plan, bundle) = parse_plan_and_bundle(config)?;
-
-    Ok(SupervisorConfig {
-        krun,
-        vm_state_dir: state_dir.to_string_lossy().into_owned(),
-        pid_file_name: None,
-        tenant_id: substrate.tenant_id,
-        audit_dir: substrate.audit_dir,
-        gateway_audit_socket: substrate.gateway_audit_socket,
-        gateway_events_socket: substrate.gateway_events_socket,
-        signing_key_path: substrate.signing_key_path,
-        plan,
-        bundle,
-        // Always carry the resolved egress policy so the bridge enforces it
-        // on the no-bundle path (deny-all included — that's the enforced
-        // default). Inert for the legacy non-bridge path (it only spawns the
-        // bridge when admitted), and superseded by `bundle` when one resolves.
-        network_policy: Some(
-            serde_json::to_value(&config.network_policy)
-                .map_err(|e| anyhow!("serialize VmStartConfig.network_policy: {e}"))?,
-        ),
-        // Only `HardFail` is
-        // implemented today. Defaulting here keeps the producer site
-        // explicit while a future change can introduce policy-driven
-        // selection.
-        bridge_restart_policy: libkrun_sys::BridgeRestartPolicy::HardFail,
-        transparent_terminator_port: Some(
-            mvm_vmm::host::egress_redirect::terminator_port_for_vm_name(&config.name),
-        ),
-        egress_relay_socket: None,
-        exclusive_image_lock: None,
-    })
-}
-
-fn supervisor_config_dump_path(state_dir: &Path) -> PathBuf {
-    state_dir.join("supervisor-config.json")
-}
-
-fn persist_supervisor_config_dump(state_dir: &Path, json: &str) -> Result<()> {
-    let path = supervisor_config_dump_path(state_dir);
-    std::fs::write(&path, json).map_err(|e| anyhow!("write {}: {e}", path.display()))
-}
-
-fn read_last_bytes_of(path: &Path, max_bytes: u64) -> std::io::Result<String> {
-    use std::io::{Read, Seek, SeekFrom};
-    let mut file = std::fs::File::open(path)?;
-    let len = file.metadata()?.len();
-    let take = max_bytes.min(len);
-    let offset = i64::try_from(take).unwrap_or(i64::MAX).saturating_neg();
-    file.seek(SeekFrom::End(offset))?;
-    let mut buf = Vec::with_capacity(take as usize);
-    file.read_to_end(&mut buf)?;
-    Ok(String::from_utf8_lossy(&buf).into_owned())
-}
-
-fn summarize_supervisor_config(path: &Path) -> String {
-    if !path.exists() {
-        return format!("supervisor-config.json: missing ({})", path.display());
-    }
-    let body = match std::fs::read_to_string(path) {
-        Ok(body) => body,
-        Err(err) => {
-            return format!(
-                "supervisor-config.json: present but unreadable ({}): {err}",
-                path.display()
-            );
-        }
-    };
-    if body.trim().is_empty() {
-        return format!(
-            "supervisor-config.json: present but empty ({})",
-            path.display()
-        );
-    }
-    let value: serde_json::Value = match serde_json::from_str(&body) {
-        Ok(value) => value,
-        Err(err) => {
-            return format!(
-                "supervisor-config.json: present but invalid json ({}): {err}",
-                path.display()
-            );
-        }
-    };
-    let kernel_path = value
-        .pointer("/krun/kernel_path")
-        .and_then(serde_json::Value::as_str)
-        .unwrap_or("<none>");
-    let rootfs_path = value
-        .pointer("/krun/rootfs_path")
-        .and_then(serde_json::Value::as_str)
-        .unwrap_or("<none>");
-    let initramfs_path = value
-        .pointer("/krun/initramfs_path")
-        .and_then(serde_json::Value::as_str)
-        .unwrap_or("<none>");
-    let networking = value
-        .pointer("/krun/networking")
-        .and_then(serde_json::Value::as_str)
-        .unwrap_or("<none>");
-    let disks = value
-        .pointer("/krun/disks")
-        .and_then(serde_json::Value::as_array)
-        .map(|entries| {
-            entries
-                .iter()
-                .filter_map(|entry| entry.get("id").and_then(serde_json::Value::as_str))
-                .collect::<Vec<_>>()
-        })
-        .unwrap_or_default();
-    let cmdline = value
-        .pointer("/krun/cmdline")
-        .and_then(serde_json::Value::as_str)
-        .unwrap_or("<none>");
-    let console_output_path = value
-        .pointer("/krun/console_output_path")
-        .and_then(serde_json::Value::as_str)
-        .unwrap_or("<none>");
-    format!(
-        "supervisor-config.json: {}\n  kernel_path={kernel_path}\n  rootfs_path={rootfs_path}\n  initramfs_path={initramfs_path}\n  networking={networking}\n  console_output_path={console_output_path}\n  disks={} [{}]\n  cmdline={cmdline}",
-        path.display(),
-        disks.len(),
-        disks.join(", ")
-    )
-}
-
-fn diagnose_log_path(label: &str, path: &Path) -> String {
-    if !path.exists() {
-        return format!("{label}: missing ({})", path.display());
-    }
-    match std::fs::metadata(path) {
-        Ok(metadata) if metadata.len() == 0 => {
-            format!("{label}: present but empty ({})", path.display())
-        }
-        Ok(_) => match read_last_bytes_of(path, 4096) {
-            Ok(tail) => format!("{label}: {}\n{}", path.display(), tail.trim_end()),
-            Err(err) => format!(
-                "{label}: present but unreadable ({}): {err}",
-                path.display()
-            ),
-        },
-        Err(err) => format!(
-            "{label}: present but unreadable ({}): {err}",
-            path.display()
-        ),
-    }
-}
-
-fn diagnose_vsock_sockets(state_dir: &Path) -> String {
-    let mut sockets = match std::fs::read_dir(state_dir) {
-        Ok(entries) => entries
-            .filter_map(|entry| {
-                let entry = entry.ok()?;
-                let file_name = entry.file_name();
-                let name = file_name.to_str()?;
-                (name.starts_with("vsock-") && name.ends_with(".sock")).then(|| name.to_string())
-            })
-            .collect::<Vec<_>>(),
-        Err(err) => {
-            return format!(
-                "vsock sockets: unreadable state dir ({}): {err}",
-                state_dir.display()
-            );
-        }
-    };
-    sockets.sort();
-    if sockets.is_empty() {
-        format!("vsock sockets: none under {}", state_dir.display())
-    } else {
-        format!(
-            "vsock sockets under {}: {}",
-            state_dir.display(),
-            sockets.join(", ")
-        )
-    }
-}
-
-fn libkrun_startup_diagnostics(state_dir: &Path) -> String {
-    let console_log = state_dir.join("console.log");
-    let supervisor_stderr = state_dir.join("supervisor.stderr.log");
-    let supervisor_lifecycle = state_dir.join("supervisor.lifecycle.log");
-    let supervisor_config = supervisor_config_dump_path(state_dir);
-    format!(
-        "libkrun vm state dir: {}\n{}\n{}\n{}\n{}\n{}",
-        state_dir.display(),
-        diagnose_vsock_sockets(state_dir),
-        diagnose_log_path("supervisor.lifecycle.log", &supervisor_lifecycle),
-        summarize_supervisor_config(&supervisor_config),
-        diagnose_log_path("supervisor.stderr.log", &supervisor_stderr),
-        diagnose_log_path("console.log", &console_log),
-    )
-}
-
-impl VmBackend for LibkrunBackend {
-    fn name(&self) -> &str {
-        "libkrun"
-    }
-
-    fn kind(&self) -> BackendKind {
-        BackendKind::Libkrun
-    }
-
-    fn capabilities(&self) -> VmCapabilities {
-        // libkrun does not support memory snapshots (same trade as
-        // Apple Container). The mvm libkrun launch path is intentionally
-        // vsock-only: the supervisor accepts only `NetworkingMode::VsockDirect`,
-        // which configures a virtio-vsock device and no net device at all, and
-        // serves egress through the host-bound vsock proxy.
-        // Pause/resume is theoretically possible but not exposed by libkrun's
-        // public C API today.
-        VmCapabilities {
-            pause_resume: false,
-            snapshots: false,
-            snapshot_capability: SnapshotCapability::DiskOnly,
-            standby_pool: true,
-            vsock: true,
-            tap_networking: false,
-            // Stronger than the field name asks for: the guest has no NIC at
-            // all, not a NIC without a route. `VsockDirect` configures a
-            // virtio-vsock device and never calls libkrun's net attach, so
-            // there is no net device in the guest's device tree to route.
-            no_routable_guest_nic: true,
-            host_vsock_proxy: true,
-            // libkrun attaches a virtio-net device and drains it, which
-            // satisfies `no_routable_guest_nic` (no upstream route) but
-            // not the L3 tunnel's stricter precondition: that mode
-            // requires the guest to have no network device at all, so
-            // `mvm0` is the only interface its stack can route to.
-            l3_vsock: false,
-            // libkrun's C API doesn't expose virtio-balloon control
-            // today; the upstream crate carries no `.balloon(...)`
-            // builder. Declared `false` until wiring lands.
-            balloon: false,
-            // libkrun's krun_add_virtiofs2/3 APIs can export a host directory
-            // as a virtio-fs share, including DAX and host-enforced read-only.
-            virtiofs_root: true,
-            // libkrun runs on macOS but the rootfs lives in a regular
-            // file, not an APFS clone-eligible volume mount; no
-            // clonefile shortcut here.
-            fs_quick_checkpoint: false,
-            // Named explicitly, not left at the all-`None` struct-update
-            // default: on Linux a cgroup can bound whatever process this
-            // backend runs — libkrun does go through the dedicated
-            // mvm-libkrun-supervisor binary, but the cgroup claim doesn't
-            // depend on that; it holds for any Linux process. libkrun's
-            // macOS 13-25 default host has no cgroup at all, so
-            // `for_backend` answers host-conditionally rather than by kind.
-            resource_controls: ResourceControls::for_backend(BackendKind::Libkrun),
-            ..VmCapabilities::default()
-        }
-    }
-
-    fn start(&self, config: &VmStartConfig) -> Result<VmId> {
-        if !libkrun_sys::is_available() {
-            bail!(
-                "libkrun is not installed on this host.\n  {}",
-                libkrun_sys::install_hint()
-            );
-        }
-        ensure_libkrun_runtime_source_supported(config)?;
-
-        // Early kernel-path check. `build_supervisor_config`
-        // re-checks too, but this keeps the existing test contract
-        // (`libkrun_start_errors_when_kernel_path_missing`) firing before
-        // `admit_overlay_aware`'s sidecar check would mask it.
-        let _ = config
-            .kernel_path
-            .as_deref()
-            .ok_or_else(|| anyhow!("libkrun backend requires a kernel path"))?;
-        let supervisor_path = resolve_supervisor_path()?;
-        let state_dir = vm_state_dir(&config.name);
-        std::fs::create_dir_all(&state_dir)
-            .map_err(|e| anyhow!("create per-VM state dir {}: {e}", state_dir.display()))?;
-
-        // Thread the build-time sidecar into per-VM runtime
-        // metadata so `mvmctl console` enforces the accessible/sealed
-        // gate on libkrun-launched VMs the same way as on the
-        // libkrun/Firecracker paths.
-        let rootfs = Path::new(&config.rootfs_path);
-        // Admission gate — refuse a rootfs that lacks the
-        // `/mvm/runtime` mount point. Fires before the supervisor
-        // spawn so a refusal leaves no PID file or krun handle behind.
-        let rootfs_dir = rootfs.parent().unwrap_or_else(|| Path::new("."));
-        mvm_vmm::host::runtime_meta::admit_runtime_overlay_contract(
-            rootfs_dir,
-            config.runtime_source_policy,
-        )?;
-        mvm_vmm::host::runtime_meta::record_from_start_config(
-            &config.name,
-            StartMode::Detached,
-            config,
-        )?;
-
-        let vcpus = u8::try_from(config.cpus.clamp(1, u32::from(u8::MAX))).unwrap_or(u8::MAX);
-        let cfg = build_supervisor_config(config, &state_dir)?;
-        // The L3 gateway is orthogonal to the substitution endpoint: a
-        // launch may need one, the other, or neither. It goes first because
-        // it must be listening before the guest boots and dials it.
-        mvm_vmm::host::netd_spawn::spawn_netd_if_needed(
-            config,
-            &state_dir,
-            mvm_core::vm_backend::BackendKind::Libkrun,
-        )?;
-        let mut endpoint_guard = spawn_libkrun_egress_endpoint_if_needed(
-            &config.name,
-            &state_dir,
-            config.tenant_id.as_deref().unwrap_or("local"),
-            &config.network_policy,
-        )?;
-        let pid_file = cfg.pid_file();
-        // Remove any stale PID file from a previous crashed supervisor
-        // so the wait-loop below can detect the new one unambiguously.
-        let _ = std::fs::remove_file(&pid_file);
-
-        let json =
-            serde_json::to_string(&cfg).map_err(|e| anyhow!("serialize SupervisorConfig: {e}"))?;
-        persist_supervisor_config_dump(&state_dir, &json)?;
-
-        ui::info(&format!(
-            "Starting libkrun VM '{}' (cpus={vcpus}, mem={}MiB) via {}...",
-            config.name,
-            config.memory_mib,
-            supervisor_path.display(),
-        ));
-
-        // Capture the guest console (hvc0 → supervisor stdout) to a
-        // per-VM log so a boot failure / kernel panic is diagnosable
-        // (mirrors firecracker.log). Best-effort: fall back to null if
-        // the file can't be created.
-        let console_log =
-            mvm_vmm::host::console_capture::open_console_capture(&vm_console_log(&config.name))
-                .map(Stdio::from)
-                .unwrap_or_else(|_| Stdio::null());
-        let supervisor_stderr = mvm_vmm::host::console_capture::open_console_capture(
-            &state_dir.join("supervisor.stderr.log"),
-        )
-        .map(Stdio::from)
-        .unwrap_or_else(|_| Stdio::inherit());
-        let mut child = Command::new(&supervisor_path)
-            .stdin(Stdio::piped())
-            .stdout(console_log)
-            .stderr(supervisor_stderr)
-            .spawn()
-            .map_err(|e| anyhow!("spawn {}: {e}", supervisor_path.display()))?;
-        child
-            .stdin
-            .take()
-            .ok_or_else(|| anyhow!("supervisor stdin was not piped"))?
-            .write_all(json.as_bytes())
-            .map_err(|e| anyhow!("pipe SupervisorConfig to supervisor stdin: {e}"))?;
-
-        // Poll for the PID file to appear; if the supervisor exits
-        // first, surface its status instead of waiting the full
-        // timeout.
-        let deadline = Instant::now() + PID_FILE_TIMEOUT;
-        loop {
-            if pid_file.exists() {
-                break;
-            }
-            if let Some(status) = child
-                .try_wait()
-                .map_err(|e| anyhow!("poll supervisor child: {e}"))?
-            {
-                bail!(
-                    "supervisor exited before writing PID file (status: {status}). \
-                     Check stderr above for libkrun errors."
-                );
-            }
-            if Instant::now() >= deadline {
-                let _ = child.kill();
-                bail!(
-                    "supervisor did not write {} within {:?}; killed",
-                    pid_file.display(),
-                    PID_FILE_TIMEOUT
-                );
-            }
-            std::thread::sleep(Duration::from_millis(50));
-        }
-
-        // The PID file means the supervisor process is up, but it binds the
-        // per-port vsock listener a beat later. Wait for the agent socket so
-        // the console attach / shell_exec that immediately follow don't race
-        // a not-yet-bound socket and report the VM "not running".
-        let agent_sock = mvm_core::config::vm_vsock_port_socket(
-            &config.name,
-            mvm_agentd::vsock::GUEST_AGENT_PORT,
-        );
-        let sock_deadline = Instant::now() + VSOCK_SOCKET_TIMEOUT;
-        while !agent_sock.exists() {
-            if let Some(status) = child
-                .try_wait()
-                .map_err(|e| anyhow!("poll supervisor child: {e}"))?
-            {
-                bail!(
-                    "supervisor exited before binding vsock socket {} (status: {status})\n{}",
-                    agent_sock.display(),
-                    libkrun_startup_diagnostics(&state_dir)
-                );
-            }
-            if Instant::now() >= sock_deadline {
-                let _ = child.kill();
-                bail!(
-                    "supervisor did not bind vsock socket {} within {:?}; killed\n{}",
-                    agent_sock.display(),
-                    VSOCK_SOCKET_TIMEOUT,
-                    libkrun_startup_diagnostics(&state_dir)
-                );
-            }
-            std::thread::sleep(Duration::from_millis(50));
-        }
-
-        // Spawn the per-VM host-services broker (+ its audit-signer) for an
-        // admitted workload so the guest can reach `host.audit.v1` over
-        // BROKER_PORT. Best-effort: an absent broker only disables that one
-        // service — the guest's emit fails, but the workload still runs and the
-        // system audit chain (written host-side, not via the broker) is intact.
-        // So a spawn failure is logged, never a launch rollback. On success the
-        // guard reaps both subprocesses if a later start step fails; it's
-        // defused once the VM is up and the `stop` path owns teardown.
-        // libkrun proxies the guest's BROKER_PORT dial straight to this per-VM
-        // socket — bound by the per-VM broker fork, or by the per-tenant daemon
-        // (the default; opt out with MVM_HOST_AGENT_DAEMON=0).
-        let broker_listen_socket =
-            mvm_core::config::vm_vsock_port_socket(&config.name, mvm_agentd::vsock::BROKER_PORT);
-        let host_services = config
-            .plan_json
-            .as_deref()
-            .map(mvm_core::plan::plan_from_admitted_json)
-            .transpose()
-            .context("parse admitted plan host services")?
-            .map_or_else(Vec::new, |plan| plan.services);
-        let mut broker_guard = if mvm_vmm::host::host_agent_spawn::host_agent_daemon_enabled() {
-            match mvm_vmm::host::host_agent_spawn::register_host_agent_services_if_admitted(
-                mvm_vmm::host::host_agent_spawn::HostAgentServicesParams {
-                    workload_id: &config.name,
-                    tenant_id: config.tenant_id.as_deref(),
-                    vm_name: &config.name,
-                    state_dir: &state_dir,
-                    broker_listen_socket: &broker_listen_socket,
-                    services: &host_services,
-                },
-            ) {
-                Ok(g) => mvm_vmm::host::host_agent_spawn::ServicesGuard::Agent(g),
-                Err(e) => {
-                    tracing::warn!(vm = %config.name, error = %e, "host-agent registration failed; host.audit.v1 unavailable for this VM");
-                    mvm_vmm::host::host_agent_spawn::ServicesGuard::None
-                }
-            }
-        } else {
-            match mvm_vmm::host::broker_services_spawn::spawn_broker_services_if_admitted(
-                mvm_vmm::host::broker_services_spawn::BrokerServicesSpawnParams {
-                    workload_id: &config.name,
-                    tenant_id: config.tenant_id.as_deref(),
-                    vm_name: &config.name,
-                    state_dir: &state_dir,
-                    broker_listen_socket: &broker_listen_socket,
-                    services: &host_services,
-                },
-            ) {
-                Ok(guard) => mvm_vmm::host::host_agent_spawn::ServicesGuard::Fork(guard),
-                Err(e) => {
-                    tracing::warn!(vm = %config.name, error = %e, "host-services broker spawn failed; host.audit.v1 unavailable for this VM");
-                    mvm_vmm::host::host_agent_spawn::ServicesGuard::None
-                }
-            }
-        };
-
-        ui::success(&format!(
-            "libkrun VM '{}' started (pid file: {}).",
-            config.name,
-            pid_file.display()
-        ));
-        endpoint_guard.defuse();
-        broker_guard.defuse();
-        Ok(VmId(config.name.clone()))
-    }
-
-    fn stop(&self, id: &VmId) -> Result<()> {
-        // Reap the per-VM substitution endpoint first (no-op when none was
-        // spawned) so a crashed VM's decrypted-secret process can't outlive the
-        // guest. Mirrors the FC `stop_vm` ordering — safe because reap is a
-        // no-op when nothing exists, even before the not-running early return.
-        mvm_vmm::host::substitution_spawn::reap_substitution_endpoint(&vm_state_dir(&id.0), &id.0);
-        mvm_vmm::host::netd_spawn::reap_netd(&vm_state_dir(&id.0));
-        // Reap the per-VM broker + audit-signer too (no-op when none spawned),
-        // so they can't outlive the guest.
-        mvm_vmm::host::broker_services_spawn::reap_broker_services(&vm_state_dir(&id.0));
-        // If this VM registered with a per-tenant host-agent daemon instead of
-        // forking a broker, deregister it + reap its audit-signer (no-op for the
-        // fork path; the daemon stays warm). The two reaps compose — each is a
-        // no-op for the other path.
-        mvm_vmm::host::host_agent_spawn::reap_host_agent_services_from_state(
-            &vm_state_dir(&id.0),
-            &id.0,
-        );
-
-        let pid_path = vm_libkrun_pid(&id.0);
-        let pid = match read_pid(&pid_path) {
-            Some(p) => p,
-            None => {
-                // No PID file: nothing to do, but tell the user so it's
-                // not silent. Matches `mvmctl stop` on a never-started
-                // VM.
-                ui::info(&format!(
-                    "libkrun VM '{}' has no PID file at {}; nothing to stop.",
-                    id.0,
-                    pid_path.display()
-                ));
-                cleanup_vsock_sockets(&vm_state_dir(&id.0));
-                return Ok(());
-            }
-        };
-
-        if !pid_alive(pid) {
-            ui::info(&format!(
-                "libkrun VM '{}' PID {pid} is not running; cleaning up state.",
-                id.0
-            ));
-            let _ = std::fs::remove_file(&pid_path);
-            cleanup_vsock_sockets(&vm_state_dir(&id.0));
-            return Ok(());
-        }
-
-        // Arm before SIGTERM so a short-lived supervisor cannot exit between
-        // signal delivery and observer registration.
-        let observer = mvm_vmm::host::process_exit::ProcessExitObserver::arm(pid).ok();
-        // SIGTERM first — gives libkrun a chance to clean up virtio-blk file
-        // descriptors. Then SIGKILL if it ignores us.
-        send_signal(pid, libc::SIGTERM);
-        let exited = mvm_vmm::host::process_exit::wait_for_pid_exit(
-            pid,
-            Instant::now() + STOP_TIMEOUT,
-            observer.as_ref(),
-        );
-        if !exited {
-            ui::info(&format!(
-                "libkrun VM '{}' PID {pid} did not exit after SIGTERM within {STOP_TIMEOUT:?}; sending SIGKILL.",
-                id.0
-            ));
-            send_signal(pid, libc::SIGKILL);
-            if !mvm_vmm::host::process_exit::wait_for_pid_exit(
-                pid,
-                Instant::now() + FORCE_KILL_TIMEOUT,
-                observer.as_ref(),
-            ) {
-                bail!(
-                    "libkrun VM '{}' PID {pid} could not be proven dead after SIGKILL; preserving {}",
-                    id.0,
-                    pid_path.display()
-                );
-            }
-        }
-
-        let _ = std::fs::remove_file(&pid_path);
-        cleanup_vsock_sockets(&vm_state_dir(&id.0));
-        ui::success(&format!("libkrun VM '{}' stopped.", id.0));
-        Ok(())
-    }
-
-    fn stop_all(&self) -> Result<()> {
-        let vms = self.list()?;
-        let mut last_err = None;
-        for vm in vms {
-            if let Err(e) = self.stop(&VmId(vm.name.clone())) {
-                tracing::warn!(name = vm.name, error = %e, "stop_all: stop failed");
-                last_err = Some(e);
-            }
-        }
-        if let Some(e) = last_err {
-            Err(e)
-        } else {
-            Ok(())
-        }
-    }
-
-    fn pause(&self, _id: &VmId) -> Result<()> {
-        bail!(
-            "pause is not supported by the libkrun backend (upstream C API does not expose vCPU pause)"
-        )
-    }
-
-    fn resume(&self, _id: &VmId) -> Result<()> {
-        bail!(
-            "resume is not supported by the libkrun backend (upstream C API does not expose vCPU pause)"
-        )
-    }
-
-    fn warm_start(
-        &self,
-        config: &VmStartConfig,
-        requested: SnapshotCapability,
-    ) -> std::result::Result<WarmStartOutcome, WarmStartError> {
-        // libkrun has no memory snapshot; a live-memory (or save/restore)
-        // request fails closed with a recovery action rather than silently
-        // cold-booting.
-        let available = SnapshotCapability::DiskOnly;
-        if !available.satisfies(requested) {
-            return Err(WarmStartError::Unsupported {
-                requested,
-                available,
-                hint: "libkrun has no live-memory snapshot — use the Firecracker (Linux) \
-                       backend for a live-memory resume, or `mvmctl up` for a cold boot"
-                    .to_string(),
-            });
-        }
-
-        // Disk-only warm-start = fast reboot from a copy-on-write disk
-        // snapshot: boot a private writable clone of the golden rootfs so the
-        // base image stays immutable across resumes (via `SnapshotUpper`).
-        let base = Path::new(&config.rootfs_path);
-        let base_dir = base.parent().ok_or_else(|| {
-            WarmStartError::Failed(format!("rootfs has no parent: {}", base.display()))
-        })?;
-        let base_name = base.file_name().and_then(|n| n.to_str()).ok_or_else(|| {
-            WarmStartError::Failed(format!("rootfs has no file name: {}", base.display()))
-        })?;
-        let upper = vm_state_dir(&config.name).join("warm-upper");
-        let snap = SnapshotUpper::new(base_dir, &upper)
-            .map_err(|e| WarmStartError::Failed(format!("open disk snapshot: {e}")))?;
-        let warm_rootfs = snap
-            .materialize_image(base_name)
-            .map_err(|e| WarmStartError::Failed(format!("materialize warm rootfs: {e}")))?;
-
-        let mut warm = config.clone();
-        warm.rootfs_path = warm_rootfs.to_string_lossy().into_owned();
-        // Disk-only reboot carries no VMGenID rotation (no live memory to
-        // reseed) — report `NotApplicable` rather than a misleading "rotated".
-        let id = self
-            .start(&warm)
-            .map_err(|e| WarmStartError::Failed(e.to_string()))?;
-        Ok(WarmStartOutcome {
-            id,
-            reseed: mvm_core::vm_backend::ReseedStatus::NotApplicable,
-        })
-    }
-
-    fn supports_standby_pool(&self) -> bool {
-        true
-    }
-
-    fn spawn_standby(
-        &self,
-        spec: &StandbySpec,
-    ) -> std::result::Result<StandbyHandle, StandbyError> {
-        let base = standby_base_config(spec);
-        // Pre-create the control-UDS parent dir 0700 (the supervisor re-binds the socket
-        // itself; this only ensures the dir exists with the right mode).
-        if let Some(parent) = Path::new(&spec.control_socket).parent() {
-            std::fs::create_dir_all(parent)
-                .map_err(|e| StandbyError::SpawnFailed(format!("create control dir: {e}")))?;
-            let _ = std::fs::set_permissions(
-                parent,
-                std::os::unix::fs::PermissionsExt::from_mode(0o700),
-            );
-        }
-        // The bin dispatches the prelaunch arm on the `prelaunch_base` wrapper key;
-        // legacy callers send a bare SupervisorConfig and are unaffected.
-        let envelope = serde_json::to_vec(&serde_json::json!({ "prelaunch_base": base }))
-            .map_err(|e| StandbyError::SpawnFailed(format!("encode base config: {e}")))?;
-
-        let bin =
-            resolve_supervisor_path().map_err(|e| StandbyError::SpawnFailed(e.to_string()))?;
-        // Capture the standby supervisor's stderr next to its control socket so a standby
-        // that dies before it's claimed leaves a diagnosable trail (it's detached, so there
-        // is no parent to inherit into).
-        let stderr = Path::new(&spec.control_socket)
-            .parent()
-            .map(|d| d.join("standby.stderr.log"))
-            .and_then(|p| std::fs::File::create(p).ok())
-            .map(Stdio::from)
-            .unwrap_or_else(Stdio::null);
-        let mut child = Command::new(bin)
-            .stdin(Stdio::piped())
-            .stdout(Stdio::null())
-            .stderr(stderr)
-            .spawn()
-            .map_err(|e| StandbyError::SpawnFailed(format!("spawn supervisor: {e}")))?;
-        child
-            .stdin
-            .take()
-            .ok_or_else(|| StandbyError::SpawnFailed("supervisor stdin unavailable".into()))?
-            .write_all(&envelope)
-            .map_err(|e| StandbyError::SpawnFailed(format!("write base config: {e}")))?;
-        // Detached: do NOT wait — the bin blocks on the control UDS until claimed (or
-        // reaped). The dropped `Child` leaves the process running (no kill-on-drop).
-        let pid = child.id();
-
-        Ok(StandbyHandle {
-            id: spec.id.clone(),
-            template_id: spec.template_id.clone(),
-            control_socket: spec.control_socket.clone(),
-            pid,
-            kernel_sha256: spec.kernel_sha256.clone(),
-            vcpus: spec.vcpus,
-            mem_mib: spec.mem_mib,
-            binding_nonce: spec.binding_nonce.clone(),
-            spawned_unix_secs: mvm_core::time::now_unix_secs(),
-            state: StandbyState::Idle,
-            image_sha256: None, // libkrun standbys are image-agnostic
-            vsock_egress: spec.vsock_egress,
-            parent_checkpoint: None, // this spawn path does not capture a checkpoint yet
-            preloaded_child_vm_name: None,
-        })
-    }
-
-    fn claim_standby(
-        &self,
-        handle: &StandbyHandle,
-        claim: &StandbyClaim,
-    ) -> std::result::Result<VmId, StandbyError> {
-        let attach = standby_attach_config(claim, handle.binding_nonce.clone())?;
-        let mut stream = UnixStream::connect(&handle.control_socket).map_err(|e| {
-            StandbyError::ClaimFailed(format!(
-                "connect control UDS {}: {e}",
-                handle.control_socket
-            ))
-        })?;
-        libkrun_sys::framing::write_json_frame_sync(&mut stream, &attach)
-            .map_err(|e| StandbyError::ClaimFailed(format!("send attach: {e}")))?;
-        // The supervisor re-verifies the attach then start_enters, writing its pid
-        // into vm_state_dir. The VM is identified by the standby id (its state lives under
-        // vms/<id>/, so stop/status/console resolve it like any cold-booted VM).
-        Ok(VmId(handle.id.clone()))
-    }
-
-    fn status(&self, id: &VmId) -> Result<VmStatus> {
-        let pid_path = vm_libkrun_pid(&id.0);
-        match read_pid(&pid_path) {
-            Some(pid) if pid_alive(pid) => Ok(VmStatus::Running),
-            _ => Ok(VmStatus::Stopped),
-        }
-    }
-
-    fn wait(&self, id: &VmId) -> Result<mvm_core::vm_backend::VmExitStatus> {
-        Ok(mvm_vmm::host::workload_wait::wait_for_workload_exit(
-            &vm_state_dir(&id.0),
-        ))
-    }
-
-    fn list(&self) -> Result<Vec<VmInfo>> {
-        let root = vms_dir();
-        let entries = match std::fs::read_dir(&root) {
-            Ok(it) => it,
-            Err(e) if e.kind() == std::io::ErrorKind::NotFound => return Ok(Vec::new()),
-            Err(e) => return Err(anyhow!("read {}: {e}", root.display())),
-        };
-        let mut vms = Vec::new();
-        for entry in entries.flatten() {
-            let path = entry.path();
-            if !path.is_dir() {
-                continue;
-            }
-            let pid_path = path.join("libkrun.pid");
-            if !pid_path.exists() {
-                // Not a libkrun-managed VM (e.g. Apple Container under
-                // the same ~/.mvm/vms/ root).
-                continue;
-            }
-            let name = match entry.file_name().into_string() {
-                Ok(s) => s,
-                Err(_) => continue, // skip non-UTF-8 dir names
-            };
-            let alive = read_pid(&pid_path).is_some_and(pid_alive);
-            // The supervisor doesn't surface cpus/memory/ports back to
-            // the parent today; a future status RPC over vsock will read
-            // these out of the live VM.
-            // Until then, `list` reports the name + running state and
-            // leaves the discoverable-from-runtime fields empty.
-            vms.push(VmInfo {
-                id: VmId(name.clone()),
-                name,
-                status: if alive {
-                    VmStatus::Running
-                } else {
-                    VmStatus::Stopped
-                },
-                guest_ip: None,
-                cpus: 0,
-                memory_mib: 0,
-                profile: None,
-                revision: None,
-                flake_ref: None,
-                ports: Vec::new(),
-            });
-        }
-        Ok(vms)
-    }
-
-    fn logs(&self, id: &VmId, _lines: u32, _hypervisor: bool) -> Result<String> {
-        bail!("libkrun logs not yet implemented for VM '{}'", id.0)
-    }
-
-    fn is_available(&self) -> Result<bool> {
-        Ok(libkrun_sys::is_available())
-    }
-
-    fn install(&self) -> Result<()> {
-        ui::info(&format!(
-            "libkrun must be installed via the host's package manager.\n  {}",
-            libkrun_sys::install_hint()
-        ));
-        // Also surface where we look for the supervisor binary so
-        // operators can pre-flight the install before hitting a
-        // mid-`mvmctl up` failure.
-        match resolve_supervisor_path() {
-            Ok(path) => ui::info(&format!("Supervisor binary: {}", path.display())),
-            Err(e) => ui::info(&format!("Supervisor binary: NOT FOUND ({e})")),
-        }
-        Ok(())
-    }
-
-    fn guest_channel_info(&self, _id: &VmId) -> Result<GuestChannelInfo> {
-        // libkrun exposes vsock as a host-side abstract socket; the
-        // guest agent listens on the shared `GUEST_AGENT_PORT` port,
-        // identical to Firecracker and Apple Container, so callers can
-        // share the same vsock client implementation across backends.
-        Ok(GuestChannelInfo::Vsock {
-            cid: 3, // standard guest CID
-            port: mvm_agentd::vsock::GUEST_AGENT_PORT,
-        })
-    }
-
-    fn security_profile(&self) -> BackendSecurityProfile {
-        // Tier 2: hardware isolation via KVM (Linux) or Hypervisor.framework
-        // (macOS). Comparable VMM TCB to Firecracker — libkrun is rust-vmm
-        // based, ~80K LOC, no Firecracker-excluded features (so it passes
-        // the "fork test"). Claim 3 (verified boot) is partial
-        // because the dm-verity pipeline currently targets Firecracker;
-        // libkrun support is a follow-up.
-        BackendSecurityProfile {
-            claims: [
-                ClaimStatus::Holds,       // 1 — host-fs isolation via KVM/HVF
-                ClaimStatus::Holds,       // 2 — uid-0 protections same as FC
-                ClaimStatus::DoesNotHold, // 3 — verified boot for libkrun rootfs not yet wired
-                ClaimStatus::Holds,       // 4 — guest agent has no do_exec in prod
-                ClaimStatus::Holds,       // 5 — vsock framing is fuzzed
-                ClaimStatus::Holds,       // 6 — image hash verification
-                ClaimStatus::Holds,       // 7 — cargo deps audited
-            ],
-            layer_coverage: LayerCoverage::all_layers(),
-            tier: "Tier 2",
-            notes: &[
-                "Hardware isolation via KVM (Linux) or Hypervisor.framework (macOS).",
-                "Comparable VMM TCB to Firecracker; passes plan 53 §\"fork test\".",
-                "Claim 3 (verified boot) is partial — dm-verity pipeline targets Firecracker today.",
-                "Supported on Linux KVM and macOS Apple Silicon; macOS Intel is not a supported local host.",
-            ],
-        }
-    }
-}
-
-// ─── helpers ───────────────────────────────────────────────────────
-
-/// Resolve the absolute path to the `mvm-libkrun-supervisor` binary. Compiled
-/// by mvmctl's build script; see [`mvm_vmm::host::aux_bin`] for the search order.
-pub fn resolve_supervisor_path() -> Result<PathBuf> {
-    mvm_vmm::host::aux_bin::resolve(&mvm_vmm::host::aux_bin::AuxBin {
-        bin: "mvm-libkrun-supervisor",
-        env_var: "MVM_LIBKRUN_SUPERVISOR_PATH",
-    })
-}
-
-pub(crate) fn read_pid(path: &Path) -> Option<libc::pid_t> {
-    let s = std::fs::read_to_string(path).ok()?;
-    s.trim().parse::<libc::pid_t>().ok()
-}
-
-pub(crate) fn pid_alive(pid: libc::pid_t) -> bool {
-    mvm_vmm::host::process_liveness::pid_is_alive(pid)
-}
-
-pub(crate) fn send_signal(pid: libc::pid_t, sig: libc::c_int) {
-    unsafe { libc::kill(pid, sig) };
-}
-
-pub(crate) fn cleanup_vsock_sockets(state_dir: &Path) {
-    let control_ports = [
-        mvm_agentd::vsock::GUEST_AGENT_PORT,
-        mvm_agentd::vsock::WORKLOAD_EXIT_PORT,
-        mvm_agentd::vsock::EGRESS_PORT,
-        mvm_agentd::vsock::BROKER_PORT,
-    ];
-    for port in control_ports
-        .into_iter()
-        .chain(mvm_agentd::vsock::dev_console_data_ports())
-    {
-        let socket = mvm_core::config::vm_vsock_port_socket_at(state_dir, port);
-        let _ = std::fs::remove_file(socket);
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use mvm_core::util::test_env::TestEnv;
-
-    #[test]
-    fn vsock_cleanup_removes_only_known_port_paths() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let agent = mvm_core::config::vm_vsock_port_socket_at(
-            temp.path(),
-            mvm_agentd::vsock::GUEST_AGENT_PORT,
-        );
-        let workload_exit = mvm_core::config::vm_vsock_port_socket_at(
-            temp.path(),
-            mvm_agentd::vsock::WORKLOAD_EXIT_PORT,
-        );
-        let unrelated = temp.path().join("unrelated.sock");
-        std::fs::create_dir_all(agent.parent().expect("socket parent")).expect("create socket dir");
-        std::fs::write(&agent, b"agent").expect("write agent path");
-        std::fs::write(&workload_exit, b"exit").expect("write exit path");
-        std::fs::write(&unrelated, b"unrelated").expect("write unrelated path");
-
-        cleanup_vsock_sockets(temp.path());
-
-        assert!(!agent.exists());
-        assert!(!workload_exit.exists());
-        assert!(unrelated.exists());
-    }
-
-    fn sample_standby_spec() -> StandbySpec {
-        StandbySpec {
-            id: "standby-x".into(),
-            template_id: None,
-            kernel_path: "/k/vmlinux".into(),
-            kernel_sha256: "a".repeat(64),
-            vcpus: 2,
-            mem_mib: 1024,
-            signing_key_path: "/keys/host-signer.ed25519".into(),
-            signer_id: "host:test".into(),
-            binding_nonce: "ab".repeat(32),
-            control_socket: "/p/standby-x/control-ab.sock".into(),
-            vm_state_dir: "/vms/standby-x".into(),
-            image_path: None,
-            image_sha256: None,
-            vsock_egress: false,
-        }
-    }
-
-    fn sample_standby_claim() -> StandbyClaim {
-        StandbyClaim {
-            start_config: None,
-            rootfs_path: "/vol/rootfs.ext4".into(),
-            tenant_id: "tenant-a".into(),
-            audit_dir: "/audit".into(),
-            gateway_audit_socket: "/audit/g.sock".into(),
-            gateway_events_socket: None,
-            plan_json: r#"{"signed":"envelope"}"#.into(),
-            bundle_json: None,
-            network_policy: mvm_core::network_policy::NetworkPolicy::deny_all(),
-        }
-    }
-
-    #[test]
-    fn standby_spec_translates_to_base_config_with_no_rootfs() {
-        let spec = sample_standby_spec();
-        let base = standby_base_config(&spec);
-        // A standby carries NO workload rootfs — it arrives in the attach.
-        assert!(base.krun.rootfs_path.is_none());
-        assert_eq!(base.binding_nonce, spec.binding_nonce);
-        assert_eq!(base.signer_id, spec.signer_id);
-        assert_eq!(base.control_socket_path, spec.control_socket);
-        assert_eq!(base.krun.vcpus, 2);
-        assert_eq!(base.krun.ram_mib, 1024);
-        // The bin dispatches the prelaunch arm on the `prelaunch_base` wrapper key.
-        let env = serde_json::json!({ "prelaunch_base": base });
-        assert!(env.get("prelaunch_base").is_some());
-    }
-
-    #[test]
-    fn standby_claim_translates_to_attach_config_echoing_nonce() {
-        let attach = standby_attach_config(&sample_standby_claim(), "ab".repeat(32)).unwrap();
-        assert_eq!(attach.binding_nonce, "ab".repeat(32));
-        assert_eq!(attach.rootfs_path, "/vol/rootfs.ext4");
-        assert_eq!(attach.tenant_id, "tenant-a");
-        assert_eq!(attach.plan, serde_json::json!({"signed": "envelope"}));
-        assert!(attach.bundle.is_none());
-    }
-
-    #[test]
-    fn standby_claim_rejects_malformed_plan_json() {
-        let mut claim = sample_standby_claim();
-        claim.plan_json = "not json".into();
-        match standby_attach_config(&claim, "ab".repeat(32)) {
-            Err(StandbyError::ClaimFailed(_)) => {}
-            other => panic!("expected ClaimFailed, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn prod_console_attachment_has_no_input() {
-        // claim 15: the backend captures the guest console
-        // to a WRITE-ONLY sink — there is no host fd from which the guest
-        // could read console input. The only interactive path is the
-        // runtime-gated agent vsock console, which a sealed prod run
-        // cannot authorize.
-        use std::io::{Read, Write};
-        let tmp = tempfile::tempdir().expect("tempdir");
-        let p = tmp.path().join("console.log");
-        let mut f = mvm_vmm::host::console_capture::open_console_capture(&p).expect("open sink");
-        f.write_all(b"boot log line\n")
-            .expect("console capture must accept output");
-        let mut buf = String::new();
-        assert!(
-            f.read_to_string(&mut buf).is_err(),
-            "console capture sink must be write-only — no guest console input path"
-        );
-    }
-
-    #[test]
-    fn libkrun_backend_name() {
-        assert_eq!(LibkrunBackend.name(), "libkrun");
-    }
-
-    #[test]
-    fn libkrun_capabilities() {
-        let caps = LibkrunBackend.capabilities();
-        assert!(caps.vsock);
-        assert!(caps.no_routable_guest_nic);
-        assert!(caps.host_vsock_proxy);
-        assert!(!caps.snapshots);
-        assert!(!caps.pause_resume);
-        assert!(!caps.tap_networking);
-        assert!(!caps.l3_vsock);
-    }
-
-    #[test]
-    fn libkrun_security_profile_is_tier_2_with_partial_claim_3() {
-        let profile = LibkrunBackend.security_profile();
-        assert_eq!(profile.tier, "Tier 2");
-        assert!(profile.layer_coverage.is_microvm());
-        // Claim 3 (verified boot) is the only partial claim; everything
-        // else holds because libkrun matches Firecracker's L2 properties.
-        assert_eq!(profile.dropped_claims(), vec![3]);
-        assert!(profile.na_claims().is_empty());
-    }
-
-    #[test]
-    fn build_supervisor_config_maps_substrate_into_supervisor_config() {
-        // When the producer threaded an admitted plan,
-        // build_supervisor_config maps the resolved AuditSubstrate fields
-        // into SupervisorConfig + parses plan_json/bundle_json as JSON.
-        let config = VmStartConfig {
-            name: "test-vm".into(),
-            rootfs_path: "/tmp/rootfs.ext4".into(),
-            kernel_path: Some("/tmp/vmlinux".into()),
-            cpus: 1,
-            memory_mib: 256,
-            tenant_id: Some("acme".into()),
-            plan_json: Some("{\"k\":\"v\"}".into()),
-            bundle_json: None,
-            ..Default::default()
-        };
-        let tmp = tempfile::tempdir().unwrap();
-        let cfg = build_supervisor_config(&config, tmp.path()).expect("build");
-        assert_eq!(cfg.tenant_id.as_deref(), Some("acme"));
-        assert!(cfg.audit_dir.is_some());
-        assert!(cfg.gateway_audit_socket.is_some());
-        assert!(cfg.gateway_events_socket.is_some());
-        assert!(cfg.signing_key_path.is_some());
-        assert!(cfg.plan.is_some());
-        assert!(cfg.bundle.is_none());
-    }
-
-    #[test]
-    fn libkrun_kernel_for_host_keeps_missing_kernel_raw_for_late_error() {
-        let (kernel, format) =
-            libkrun_kernel_for_host("/tmp/mvm-test-missing-vmlinux").expect("no fs read");
-        assert_eq!(kernel, "/tmp/mvm-test-missing-vmlinux");
-        assert_eq!(format, KernelFormat::Raw);
-    }
-
-    #[test]
-    fn libkrun_kernel_for_host_marks_existing_x86_elf_as_elf() {
-        let tmp = tempfile::tempdir().unwrap();
-        let kernel_path = tmp.path().join("vmlinux");
-        std::fs::write(&kernel_path, b"\x7fELFstub").unwrap();
-        let (kernel, format) =
-            libkrun_kernel_for_host(kernel_path.to_str().expect("utf8 temp path")).unwrap();
-        assert_eq!(kernel, kernel_path.to_string_lossy());
-        if cfg!(target_arch = "x86_64") {
-            assert_eq!(format, KernelFormat::Elf);
-        } else {
-            assert_eq!(format, KernelFormat::Raw);
-        }
-    }
-
-    #[test]
-    fn libkrun_build_supervisor_config_uses_verity_shape_and_runtime_overlay_tokens() {
-        let dir = tempfile::tempdir().unwrap();
-        let rootfs = dir.path().join("rootfs.ext4");
-        let verity = dir.path().join("rootfs.verity");
-        let initrd = dir.path().join("rootfs.initrd");
-        std::fs::write(&rootfs, b"rootfs").unwrap();
-        std::fs::write(&verity, b"verity").unwrap();
-        std::fs::write(&initrd, b"initrd").unwrap();
-
-        let config = VmStartConfig {
-            name: "verity".into(),
-            rootfs_path: rootfs.display().to_string(),
-            kernel_path: Some("/tmp/vmlinux".into()),
-            cpus: 1,
-            memory_mib: 256,
-            verity_path: Some(verity.display().to_string()),
-            roothash: Some("a".repeat(64)),
-            runtime_overlay_path: Some("/tmp/runtime.ext4".into()),
-            runtime_overlay_verity_path: Some("/tmp/runtime.verity".into()),
-            runtime_overlay_roothash: Some("b".repeat(64)),
-            runtime_source_policy: mvm_core::vm_backend::RuntimeSourcePolicy::RequiredOverlay,
-            ..Default::default()
-        };
-        let cfg = build_supervisor_config(&config, dir.path()).expect("build");
-        let cmdline = cfg.krun.kernel_cmdline.as_deref().expect("cmdline is set");
-        assert!(!cmdline.contains("root=/dev/vda"));
-        assert!(!cmdline.contains("init=/init"));
-        assert!(cmdline.contains("mvm.roothash="));
-        assert!(cmdline.contains("mvm.data=/dev/vdb"));
-        assert!(cmdline.contains("mvm.hash=/dev/vdc"));
-        assert!(cmdline.contains("mvm.runtime_roothash="));
-        assert!(cmdline.contains("mvm.runtime_data=/dev/vdd"));
-        assert!(cmdline.contains("mvm.runtime_hash=/dev/vde"));
-        assert!(cmdline.contains("mvm.runtime_source_policy=required_overlay"));
-        assert_eq!(
-            cfg.krun.initramfs_path.as_deref(),
-            Some(initrd.to_string_lossy().as_ref())
-        );
-        assert!(cfg.krun.rootfs_path.is_none());
-        assert_eq!(
-            cfg.krun.console_output_path.as_deref(),
-            Some(vm_console_log("verity").to_string_lossy().as_ref())
-        );
-    }
-
-    #[test]
-    fn libkrun_verity_cmdline_args_shift_devices_for_initrd_boot() {
-        let config = VmStartConfig {
-            rootfs_path: "/tmp/rootfs.ext4".into(),
-            verity_path: Some("/tmp/rootfs.verity".into()),
-            roothash: Some("a".repeat(64)),
-            runtime_overlay_path: Some("/tmp/runtime.ext4".into()),
-            runtime_overlay_verity_path: Some("/tmp/runtime.verity".into()),
-            runtime_overlay_roothash: Some("b".repeat(64)),
-            ..Default::default()
-        };
-
-        let args = libkrun_verity_cmdline_args(&config).expect("verity cmdline args");
-        assert!(args.contains("mvm.data=/dev/vdb"), "got: {args}");
-        assert!(args.contains("mvm.hash=/dev/vdc"), "got: {args}");
-        assert!(args.contains("mvm.runtime_data=/dev/vdd"), "got: {args}");
-        assert!(args.contains("mvm.runtime_hash=/dev/vde"), "got: {args}");
-    }
-
-    #[test]
-    fn libkrun_build_supervisor_config_uses_read_only_verity_disk_layout_with_overlay() {
-        let dir = tempfile::tempdir().unwrap();
-        let rootfs = dir.path().join("rootfs.ext4");
-        let verity = dir.path().join("rootfs.verity");
-        let initrd = dir.path().join("rootfs.initrd");
-        std::fs::write(&rootfs, b"rootfs").unwrap();
-        std::fs::write(&verity, b"verity").unwrap();
-        std::fs::write(&initrd, b"initrd").unwrap();
-
-        let config = VmStartConfig {
-            name: "verity".into(),
-            rootfs_path: rootfs.display().to_string(),
-            kernel_path: Some("/tmp/vmlinux".into()),
-            cpus: 1,
-            memory_mib: 256,
-            verity_path: Some(verity.display().to_string()),
-            roothash: Some("a".repeat(64)),
-            runtime_overlay_path: Some("/tmp/runtime.ext4".into()),
-            runtime_overlay_verity_path: Some("/tmp/runtime.verity".into()),
-            runtime_overlay_roothash: Some("b".repeat(64)),
-            ..Default::default()
-        };
-        let cfg = build_supervisor_config(&config, dir.path()).expect("build");
-        assert_eq!(cfg.krun.extra_disks.len(), 4);
-        assert!(cfg.krun.extra_disks.iter().all(|disk| disk.read_only));
-        assert_eq!(cfg.krun.extra_disks[0].path, rootfs.to_string_lossy());
-        assert_eq!(cfg.krun.extra_disks[1].path, verity.to_string_lossy());
-        assert_eq!(cfg.krun.extra_disks[2].path, "/tmp/runtime.ext4");
-        assert_eq!(cfg.krun.extra_disks[3].path, "/tmp/runtime.verity");
-    }
-
-    #[test]
-    fn libkrun_non_verity_boot_attaches_runtime_overlay_as_first_extra_disk() {
-        let dir = tempfile::tempdir().unwrap();
-        let rootfs = dir.path().join("rootfs.ext4");
-        std::fs::write(&rootfs, b"rootfs").unwrap();
-
-        // A plain dev rootfs (no verity) with a resolved overlay triple: the
-        // rootfs rides /dev/vda via `rootfs_path`, so the overlay is the first
-        // extra disk (=> /dev/vdb), read-only, matching the cmdline token.
-        let config = VmStartConfig {
-            name: "dev".into(),
-            rootfs_path: rootfs.display().to_string(),
-            kernel_path: Some("/tmp/vmlinux".into()),
-            cpus: 1,
-            memory_mib: 256,
-            runtime_overlay_path: Some("/tmp/runtime.ext4".into()),
-            runtime_overlay_verity_path: Some("/tmp/runtime.verity".into()),
-            runtime_overlay_roothash: Some("b".repeat(64)),
-            runtime_source_policy: mvm_core::vm_backend::RuntimeSourcePolicy::PreferOverlay,
-            ..Default::default()
-        };
-        let cfg = build_supervisor_config(&config, dir.path()).expect("build");
-        assert_eq!(
-            cfg.krun.rootfs_path.as_deref(),
-            Some(rootfs.to_string_lossy().as_ref())
-        );
-        assert_eq!(cfg.krun.extra_disks.len(), 1);
-        assert_eq!(cfg.krun.extra_disks[0].path, "/tmp/runtime.ext4");
-        assert!(cfg.krun.extra_disks[0].read_only);
-        let cmdline = cfg.krun.kernel_cmdline.as_deref().expect("cmdline");
-        assert!(cmdline.contains("root=/dev/vda"), "got: {cmdline}");
-        assert!(
-            cmdline.contains("mvm.runtime_data=/dev/vdb"),
-            "got: {cmdline}"
-        );
-        assert!(!cmdline.contains("mvm.runtime_hash="), "got: {cmdline}");
-    }
-
-    #[test]
-    fn required_overlay_libkrun_support_rejects_missing_overlay_artifacts() {
-        let dir = tempfile::tempdir().unwrap();
-        let rootfs = dir.path().join("rootfs.ext4");
-        let verity = dir.path().join("rootfs.verity");
-        let initrd = dir.path().join("rootfs.initrd");
-        std::fs::write(&rootfs, b"rootfs").unwrap();
-        std::fs::write(&verity, b"verity").unwrap();
-        std::fs::write(&initrd, b"initrd").unwrap();
-
-        let config = VmStartConfig {
-            rootfs_path: rootfs.display().to_string(),
-            verity_path: Some(verity.display().to_string()),
-            roothash: Some("a".repeat(64)),
-            runtime_source_policy: mvm_core::vm_backend::RuntimeSourcePolicy::RequiredOverlay,
-            ..Default::default()
-        };
-        let err = ensure_libkrun_runtime_source_supported(&config).unwrap_err();
-        assert!(err.to_string().contains("runtime overlay artifact triple"));
-    }
-
-    #[test]
-    fn required_overlay_libkrun_support_rejects_missing_initrd() {
-        let dir = tempfile::tempdir().unwrap();
-        let rootfs = dir.path().join("rootfs.ext4");
-        let verity = dir.path().join("rootfs.verity");
-        std::fs::write(&rootfs, b"rootfs").unwrap();
-        std::fs::write(&verity, b"verity").unwrap();
-
-        let config = VmStartConfig {
-            rootfs_path: rootfs.display().to_string(),
-            verity_path: Some(verity.display().to_string()),
-            roothash: Some("a".repeat(64)),
-            runtime_overlay_path: Some("/tmp/runtime.ext4".into()),
-            runtime_overlay_verity_path: Some("/tmp/runtime.verity".into()),
-            runtime_overlay_roothash: Some("b".repeat(64)),
-            runtime_source_policy: mvm_core::vm_backend::RuntimeSourcePolicy::RequiredOverlay,
-            ..Default::default()
-        };
-        let err = ensure_libkrun_runtime_source_supported(&config).unwrap_err();
-        assert!(err.to_string().contains("rootfs.initrd"));
-    }
-
-    #[test]
-    fn required_overlay_libkrun_support_accepts_non_verity_block_overlay() {
-        // A non-verity dev boot carrying the resolved overlay triple is served by
-        // the plain read-only /dev/vdb mount — no verity metadata or initrd.
-        let config = VmStartConfig {
-            rootfs_path: "/tmp/rootfs.ext4".into(),
-            runtime_overlay_path: Some("/tmp/runtime.ext4".into()),
-            runtime_overlay_verity_path: Some("/tmp/runtime.verity".into()),
-            runtime_overlay_roothash: Some("b".repeat(64)),
-            runtime_source_policy: mvm_core::vm_backend::RuntimeSourcePolicy::RequiredOverlay,
-            ..Default::default()
-        };
-        assert!(ensure_libkrun_runtime_source_supported(&config).is_ok());
-    }
-
-    /// claim 1 (host-fs isolation): libkrun can't enforce a read-only virtio-fs
-    /// share at the hypervisor, so it refuses one rather than make a
-    /// false ro promise (a compromised guest could remount it rw). A
-    /// read-write share, or a read-only disk image, is fine.
-    #[test]
-    fn libkrun_refuses_read_only_virtiofs_share() {
-        use mvm_core::vm_backend::{VmVolume, VmVolumeKind};
-        let tmp = tempfile::tempdir().unwrap();
-        let base = VmStartConfig {
-            name: "ro-fs".into(),
-            rootfs_path: "/tmp/rootfs.ext4".into(),
-            kernel_path: Some("/tmp/vmlinux".into()),
-            cpus: 1,
-            memory_mib: 256,
-            ..Default::default()
-        };
-
-        // read-only dir share → refused.
-        let mut ro = base.clone();
-        ro.volumes = vec![VmVolume {
-            host: "/h/src".into(),
-            guest: "/data".into(),
-            read_only: true,
-            kind: VmVolumeKind::DirShare,
-            ..Default::default()
-        }];
-        let err = build_supervisor_config(&ro, tmp.path())
-            .unwrap_err()
-            .to_string();
-        assert!(err.contains("read-only virtio-fs"), "got: {err}");
-
-        // read-write dir share → allowed.
-        let mut rw = base.clone();
-        rw.volumes = vec![VmVolume {
-            host: "/h/src".into(),
-            guest: "/data".into(),
-            read_only: false,
-            kind: VmVolumeKind::DirShare,
-            ..Default::default()
-        }];
-        assert!(build_supervisor_config(&rw, tmp.path()).is_ok());
-
-        // read-only DISK image → allowed (hypervisor-enforced ro).
-        let mut ro_disk = base;
-        ro_disk.volumes = vec![VmVolume {
-            host: "/h/data.img".into(),
-            guest: "/data".into(),
-            size: "1G".into(),
-            read_only: true,
-            kind: VmVolumeKind::Disk,
-            ..Default::default()
-        }];
-        assert!(build_supervisor_config(&ro_disk, tmp.path()).is_ok());
-    }
-
-    #[test]
-    fn build_supervisor_config_dev_console_pre_opens_console_data_ports() {
-        // A dev-accessible managed machine must pre-open the console data
-        // range; libkrun binds one host UDS listener per registered port, so
-        // the agent's dynamic data port is otherwise unreachable on attach.
-        let config = VmStartConfig {
-            name: "console".into(),
-            rootfs_path: "/tmp/rootfs.ext4".into(),
-            kernel_path: Some("/tmp/vmlinux".into()),
-            cpus: 1,
-            memory_mib: 256,
-            dev_console: true,
-            ..Default::default()
-        };
-        let tmp = tempfile::tempdir().unwrap();
-        let cfg = build_supervisor_config(&config, tmp.path()).expect("build");
-        assert!(
-            cfg.krun
-                .vsock_ports
-                .contains(&mvm_agentd::vsock::GUEST_AGENT_PORT)
-        );
-        assert!(
-            cfg.krun
-                .vsock_ports
-                .contains(&(mvm_agentd::vsock::CONSOLE_PORT_BASE + 1)),
-            "first console data port must be registered when dev_console is set"
-        );
-        assert!(
-            cfg.krun.vsock_ports.contains(
-                &(mvm_agentd::vsock::CONSOLE_PORT_BASE
-                    + mvm_agentd::vsock::DEV_CONSOLE_DATA_PORT_COUNT)
-            ),
-            "last console data port must be registered when dev_console is set"
-        );
-
-        // The default (non-dev) boot keeps the tight agent-only port set.
-        let plain = VmStartConfig {
-            name: "plain".into(),
-            rootfs_path: "/tmp/rootfs.ext4".into(),
-            kernel_path: Some("/tmp/vmlinux".into()),
-            cpus: 1,
-            memory_mib: 256,
-            ..Default::default()
-        };
-        let cfg_plain = build_supervisor_config(&plain, tmp.path()).expect("build");
-        assert!(
-            !cfg_plain
-                .krun
-                .vsock_ports
-                .contains(&(mvm_agentd::vsock::CONSOLE_PORT_BASE + 1)),
-            "console range must NOT be registered when dev_console is unset"
-        );
-    }
-
-    #[test]
-    fn build_supervisor_config_no_tenant_keeps_substrate_none() {
-        // No admission ⇒ all five substrate fields
-        // None ⇒ supervisor takes the legacy `run_supervisor` path.
-        let config = VmStartConfig {
-            name: "dev-vm".into(),
-            rootfs_path: "/tmp/rootfs.ext4".into(),
-            kernel_path: Some("/tmp/vmlinux".into()),
-            cpus: 1,
-            memory_mib: 256,
-            ..Default::default()
-        };
-        let tmp = tempfile::tempdir().unwrap();
-        let cfg = build_supervisor_config(&config, tmp.path()).expect("build");
-        assert!(cfg.tenant_id.is_none());
-        assert!(cfg.audit_dir.is_none());
-        assert!(cfg.gateway_audit_socket.is_none());
-        assert!(cfg.gateway_events_socket.is_none());
-        assert!(cfg.signing_key_path.is_none());
-        assert!(cfg.plan.is_none());
-        assert!(cfg.bundle.is_none());
-    }
-
-    #[test]
-    fn build_supervisor_config_refuses_unsafe_tenant() {
-        // Defense-in-depth: tenant_id passes through
-        // the DNS-label allowlist in audit_substrate; build_supervisor_config
-        // propagates the error.
-        let config = VmStartConfig {
-            name: "evil-vm".into(),
-            rootfs_path: "/tmp/rootfs.ext4".into(),
-            kernel_path: Some("/tmp/vmlinux".into()),
-            cpus: 1,
-            memory_mib: 256,
-            tenant_id: Some("../escape".into()),
-            ..Default::default()
-        };
-        let tmp = tempfile::tempdir().unwrap();
-        assert!(build_supervisor_config(&config, tmp.path()).is_err());
-    }
-
-    #[test]
-    fn libkrun_install_message_is_actionable() {
-        LibkrunBackend.install().expect("install hint never errors");
-        let hint = libkrun_sys::install_hint();
-        assert!(!hint.is_empty());
-    }
-
-    #[test]
-    fn libkrun_start_errors_when_kernel_path_missing() {
-        let config = VmStartConfig {
-            name: "libkrun-test".to_string(),
-            rootfs_path: "/tmp/rootfs.ext4".to_string(),
-            ..Default::default()
-        };
-        // No kernel_path set → start should fail with a precise message
-        // before attempting to spawn the supervisor.
-        let err = LibkrunBackend
-            .start(&config)
-            .expect_err("expected failure without kernel_path");
-        let msg = err.to_string();
-        assert!(
-            msg.contains("kernel path") || msg.contains("not installed"),
-            "unexpected error message: {msg}"
-        );
-    }
-
-    /// A VM with no PID file is `Stopped`. Mirrors what `mvmctl status
-    /// <name>` reports for a never-started VM.
-    #[test]
-    fn libkrun_status_is_stopped_when_no_pid_file() {
-        let status = LibkrunBackend
-            .status(&VmId("never-started-vm".to_string()))
-            .expect("status should not error");
-        assert_eq!(status, VmStatus::Stopped);
-    }
-
-    #[test]
-    fn stop_reaps_a_supervisor_through_the_shared_exit_wait() {
-        with_env(|| {
-            let temp = tempfile::tempdir().expect("tempdir");
-            let mut env = TestEnv::new();
-            env.set("MVM_HOME", temp.path());
-
-            let vm_name = "libkrun-event-stop";
-            let mut child = std::process::Command::new("sleep")
-                .arg("30")
-                .stdin(std::process::Stdio::null())
-                .stdout(std::process::Stdio::null())
-                .stderr(std::process::Stdio::null())
-                .spawn()
-                .expect("spawn supervisor fixture");
-            let pid_path = vm_libkrun_pid(vm_name);
-            std::fs::create_dir_all(pid_path.parent().expect("pid parent")).expect("state dir");
-            std::fs::write(&pid_path, child.id().to_string()).expect("write pid marker");
-
-            LibkrunBackend
-                .stop(&VmId(vm_name.to_string()))
-                .expect("stop should prove the fixture exited");
-            let _ = child.wait();
-
-            assert!(
-                !pid_path.exists(),
-                "verified exit should remove the pid marker"
-            );
-            assert!(!pid_alive(child.id() as libc::pid_t));
-        });
-    }
-
-    /// Serialise env-var mutations across tests in this module —
-    /// cargo runs tests in parallel by default, and `std::env::set_var`
-    /// is process-global. Without this, a sibling test can flip
-    /// `HOME` (or `MVM_LIBKRUN_SUPERVISOR_PATH`) mid-call and corrupt
-    /// our assertions.
-    fn with_env<F: FnOnce()>(body: F) {
-        let _guard = mvm_vmm::host::runtime_meta::HOME_TEST_LOCK
-            .lock()
-            .unwrap_or_else(|p| p.into_inner());
-        body();
-    }
-
-    /// `list` skips directories under `~/.mvm/vms/` that don't carry a
-    /// `libkrun.pid` (e.g. Apple-Container-managed VMs sharing the
-    /// same root). Returns an empty list when the root doesn't exist.
-    #[test]
-    fn libkrun_list_returns_empty_when_no_vms_dir() {
-        with_env(|| {
-            // Point HOME at a fresh temp dir so we get a known-empty
-            // ~/.mvm/vms/ layout.
-            let temp = tempfile::tempdir().expect("tempdir");
-            let mut env = TestEnv::new();
-            env.set("HOME", temp.path());
-            let result = LibkrunBackend.list();
-            let vms = result.expect("list should not error on missing root");
-            assert!(vms.is_empty(), "expected empty list, got {vms:?}");
-        });
-    }
-
-    #[test]
-    fn resolve_supervisor_path_honors_env_override() {
-        with_env(|| {
-            let temp = tempfile::NamedTempFile::new().expect("tempfile");
-            let mut env = TestEnv::new();
-            env.set("MVM_LIBKRUN_SUPERVISOR_PATH", temp.path());
-            let result = resolve_supervisor_path();
-            let path = result.expect("env override resolves");
-            assert_eq!(path, temp.path());
-        });
-    }
-
-    #[test]
-    fn resolve_supervisor_path_rejects_missing_env_target() {
-        with_env(|| {
-            let mut env = TestEnv::new();
-            env.set(
-                "MVM_LIBKRUN_SUPERVISOR_PATH",
-                "/definitely/does/not/exist/mvm-libkrun-supervisor",
-            );
-            let result = resolve_supervisor_path();
-            let err = result.expect_err("expected missing-file error");
-            assert!(err.to_string().contains("not a file"));
-        });
-    }
-
-    #[test]
-    fn build_supervisor_config_registers_control_port() {
-        // The workload exit-code control port must appear in
-        // host_listen_ports (not vsock_ports) so the supervisor owns the
-        // listener socket and the guest connects as a client.
-        let config = VmStartConfig {
-            name: "ctrl-port-test".into(),
-            rootfs_path: "/tmp/rootfs.ext4".into(),
-            kernel_path: Some("/tmp/vmlinux".into()),
-            cpus: 1,
-            memory_mib: 256,
-            ..Default::default()
-        };
-        let tmp = tempfile::tempdir().unwrap();
-        let cfg = build_supervisor_config(&config, tmp.path()).expect("build");
-        assert!(
-            cfg.krun
-                .host_listen_ports
-                .contains(&mvm_agentd::vsock::WORKLOAD_EXIT_PORT),
-            "WORKLOAD_EXIT_PORT must be in host_listen_ports"
-        );
-        // Must not be double-registered as a guest-facing vsock port.
-        assert!(
-            !cfg.krun
-                .vsock_ports
-                .contains(&mvm_agentd::vsock::WORKLOAD_EXIT_PORT),
-            "WORKLOAD_EXIT_PORT must not appear in vsock_ports"
-        );
-    }
-
-    #[test]
-    fn build_supervisor_config_registers_substitution_port() {
-        // The egress substitution port is a host-listen port (the endpoint
-        // binds the UDS, the guest dials it) — registered unconditionally so
-        // libkrun proxies the guest's connect. Fail-closed: nothing binds the
-        // UDS unless the plan carries secrets.
-        let config = VmStartConfig {
-            name: "subst-port-test".into(),
-            rootfs_path: "/tmp/rootfs.ext4".into(),
-            kernel_path: Some("/tmp/vmlinux".into()),
-            cpus: 1,
-            memory_mib: 256,
-            ..Default::default()
-        };
-        let tmp = tempfile::tempdir().unwrap();
-        let cfg = build_supervisor_config(&config, tmp.path()).expect("build");
-        assert!(
-            cfg.krun
-                .host_listen_ports
-                .contains(&mvm_agentd::vsock::EGRESS_PORT),
-            "EGRESS_PORT must be in host_listen_ports"
-        );
-        assert!(
-            !cfg.krun
-                .vsock_ports
-                .contains(&mvm_agentd::vsock::EGRESS_PORT),
-            "EGRESS_PORT must not appear in vsock_ports"
-        );
-    }
-
-    #[test]
-    fn build_supervisor_config_registers_broker_port() {
-        // The host-services broker port is a host-listen port (the broker
-        // subprocess binds the UDS, the guest dials it) — registered
-        // unconditionally so libkrun proxies the guest's connect. Fail-closed:
-        // nothing binds the UDS until the per-VM broker subprocess is spawned,
-        // so a stray guest dial gets ECONNREFUSED.
-        let config = VmStartConfig {
-            name: "broker-port-test".into(),
-            rootfs_path: "/tmp/rootfs.ext4".into(),
-            kernel_path: Some("/tmp/vmlinux".into()),
-            cpus: 1,
-            memory_mib: 256,
-            ..Default::default()
-        };
-        let tmp = tempfile::tempdir().unwrap();
-        let cfg = build_supervisor_config(&config, tmp.path()).expect("build");
-        assert!(
-            cfg.krun
-                .host_listen_ports
-                .contains(&mvm_agentd::vsock::BROKER_PORT),
-            "BROKER_PORT must be in host_listen_ports"
-        );
-        assert!(
-            !cfg.krun
-                .vsock_ports
-                .contains(&mvm_agentd::vsock::BROKER_PORT),
-            "BROKER_PORT must not appear in vsock_ports"
-        );
-    }
-
-    #[test]
-    fn persist_supervisor_config_dump_writes_workload_json_dump() {
-        let dir = tempfile::tempdir().unwrap();
-        let cfg = SupervisorConfig {
-            krun: KrunContext::new("vm-1", "/k", "/r"),
-            vm_state_dir: "/run/state/vm-1".into(),
-            pid_file_name: None,
-            tenant_id: None,
-            audit_dir: None,
-            gateway_audit_socket: None,
-            gateway_events_socket: None,
-            signing_key_path: None,
-            plan: None,
-            bundle: None,
-            network_policy: None,
-            bridge_restart_policy: libkrun_sys::BridgeRestartPolicy::HardFail,
-            transparent_terminator_port: None,
-            egress_relay_socket: None,
-            exclusive_image_lock: None,
-        };
-        let json = serde_json::to_string(&cfg).expect("serialize");
-        persist_supervisor_config_dump(dir.path(), &json).expect("persist supervisor config");
-
-        let path = supervisor_config_dump_path(dir.path());
-        assert_eq!(path, dir.path().join("supervisor-config.json"));
-        let contents = std::fs::read_to_string(&path).expect("read persisted config");
-        assert_eq!(contents, json);
-        let value: serde_json::Value =
-            serde_json::from_str(&contents).expect("persisted config stays valid json");
-        assert_eq!(value["vm_state_dir"], "/run/state/vm-1");
-    }
-
-    #[test]
-    fn libkrun_startup_diagnostics_reports_vsock_and_log_state() {
-        let dir = tempfile::tempdir().unwrap();
-        std::fs::write(
-            dir.path().join("supervisor.lifecycle.log"),
-            "dispatch_route: direct_vsock\n",
-        )
-        .expect("write lifecycle");
-        std::fs::write(
-            dir.path().join("supervisor-config.json"),
-            r#"{"krun":{"kernel_path":"/kernel","rootfs_path":null,"initramfs_path":"/initrd","networking":"vsock_direct","console_output_path":"/state/console.log","disks":[{"id":"root"},{"id":"runtime"}],"cmdline":"console=hvc0 mvm.runtime_source_policy=required_overlay"}}"#,
-        )
-            .expect("write config");
-        std::fs::write(dir.path().join("console.log"), "").expect("write empty console log");
-        std::fs::write(dir.path().join("vsock-5251.sock"), b"")
-            .expect("write placeholder socket path");
-        std::fs::write(dir.path().join("vsock-5253.sock"), b"")
-            .expect("write placeholder socket path");
-
-        let body = libkrun_startup_diagnostics(dir.path());
-        assert!(body.contains("libkrun vm state dir:"));
-        assert!(body.contains("vsock-5251.sock"));
-        assert!(body.contains("vsock-5253.sock"));
-        assert!(body.contains("dispatch_route: direct_vsock"));
-        assert!(body.contains("kernel_path=/kernel"));
-        assert!(body.contains("initramfs_path=/initrd"));
-        assert!(body.contains("networking=vsock_direct"));
-        assert!(body.contains("console_output_path=/state/console.log"));
-        assert!(body.contains("disks=2 [root, runtime]"));
-        assert!(body.contains("mvm.runtime_source_policy=required_overlay"));
-        assert!(body.contains("console.log: present but empty"));
-    }
-
-    /// No `plan.json` plus deny-all policy ⇒ no endpoint is spawned and the
-    /// returned guard is defused (its Drop is a no-op, so an early return can't
-    /// reap a process that was never started).
-    #[test]
-    fn libkrun_substitution_not_spawned_when_no_secrets_and_no_egress() {
-        let tmp = tempfile::tempdir().unwrap();
-        let guard = spawn_libkrun_egress_endpoint_if_needed(
-            "no-secrets-vm",
-            tmp.path(),
-            "local",
-            &mvm_core::network_policy::NetworkPolicy::deny_all(),
-        )
-        .expect("deny-all no-secrets path must succeed");
-        assert!(
-            guard.vm_name.is_none(),
-            "a secret-free plan must yield a defused (no-op) guard"
-        );
-        // No pidfile was written (nothing spawned).
-        assert!(
-            !tmp.path()
-                .join(mvm_vmm::host::substitution_spawn::SUBST_PID_FILE)
-                .exists()
-        );
-    }
-
-    #[test]
-    fn vsock_egress_cmdline_token_only_when_policy_allows_egress() {
-        let dir = tempfile::tempdir().unwrap();
-        let deny_all = VmStartConfig::default();
-        assert_eq!(vsock_egress_cmdline_token(&deny_all, dir.path()), None);
-
-        let allow_egress = VmStartConfig {
-            network_policy: mvm_core::network_policy::NetworkPolicy::preset(
-                mvm_core::network_policy::NetworkPreset::Dev,
-            ),
-            ..Default::default()
-        };
-        assert_eq!(
-            vsock_egress_cmdline_token(&allow_egress, dir.path()).as_deref(),
-            Some("mvm.vsock_egress=1")
-        );
-    }
-
-    fn seed_grant_sidecar_and_key(vm_name: &str) {
-        use mvm_core::plan::{Nonce, VerbGrant, VerbId};
-        use mvm_core::protocol::vm_backend::VerbGrantEnvelope;
-
-        let state_dir = mvm_core::config::vm_state_dir(vm_name);
-        std::fs::create_dir_all(&state_dir).unwrap();
-        let nonce = Nonce::from_bytes([3u8; 16]);
-        let not_after = mvm_core::time::parse_iso8601("2099-01-01T00:00:00Z").unwrap();
-        let envelope = VerbGrantEnvelope {
-            pubkey_hex: "cc".repeat(32),
-            plan_nonce_hex: nonce.as_hex().to_string(),
-            predecessor_session_id: None,
-            predecessor_plan_nonce_hex: None,
-            grant: VerbGrant {
-                session_id: vm_name.to_string(),
-                plan_nonce: nonce,
-                not_after,
-                verbs: vec![VerbId::new("run-entrypoint").unwrap()],
-                sig: vec![0u8; 64],
-            },
-        };
-        std::fs::write(
-            state_dir.join("verb-grant.json"),
-            serde_json::to_vec(&envelope).unwrap(),
-        )
-        .unwrap();
-        let keys_dir = mvm_core::config::mvm_keys_dir();
-        std::fs::create_dir_all(&keys_dir).unwrap();
-        std::fs::write(keys_dir.join("host-signer.pub"), [0xEEu8; 32]).unwrap();
-    }
-
-    /// The libkrun supervisor cmdline must carry all three plan-bound grant
-    /// tokens — the verb grant envelope, the enforcement assertion, and the
-    /// host-signer trust anchor — mirroring the HVF `grant_tokens` parity test.
-    /// A vsock-only guest has no config drive, so all three ride the cmdline;
-    /// an auto-fallback HVF→libkrun must not land on a backend that drops them.
-    #[test]
-    fn build_supervisor_config_cmdline_carries_all_three_grant_tokens() {
-        let _guard = mvm_vmm::host::runtime_meta::HOME_TEST_LOCK
-            .lock()
-            .unwrap_or_else(|p| p.into_inner());
-        let dir = tempfile::tempdir().unwrap();
-        let mut env = TestEnv::new();
-        env.set("MVM_HOME", dir.path());
-
-        let vm_name = "libkrun-grant-parity";
-        seed_grant_sidecar_and_key(vm_name);
-
-        let config = VmStartConfig {
-            name: vm_name.into(),
-            rootfs_path: "/tmp/rootfs.ext4".into(),
-            kernel_path: Some("/tmp/vmlinux".into()),
-            cpus: 1,
-            memory_mib: 256,
-            ..Default::default()
-        };
-        let state_dir = mvm_core::config::vm_state_dir(vm_name);
-        let cfg = build_supervisor_config(&config, &state_dir).expect("build");
-        let cmdline = cfg
-            .krun
-            .kernel_cmdline
-            .as_deref()
-            .expect("supervisor cmdline present");
-        assert!(
-            cmdline.contains("mvm.verb_grant="),
-            "verb grant token missing: {cmdline}"
-        );
-        assert!(
-            cmdline.contains(
-                &mvm_vmm::host::egress_bridge::require_grant_cmdline_token(vm_name).unwrap()
-            ),
-            "require_grant token missing: {cmdline}"
-        );
-        assert!(
-            cmdline.contains("mvm.host_signer_pub="),
-            "host_signer_pub trust anchor missing: {cmdline}"
-        );
-    }
-
-    /// `capabilities()` is a declaration table that callers route on, and
-    /// nothing asserted its contents: deleting any of seven fields from
-    /// the struct literal silently fell back to the `Default`, and every
-    /// one of those deletions survived.
-    ///
-    /// `tap_networking` and `no_routable_guest_nic` are the two that
-    /// matter beyond feature routing. A libkrun workload guest has no net
-    /// device at all — the supervisor accepts only `VsockDirect`, which
-    /// never calls libkrun's net attach — so the pair must read
-    /// "no TAP, no routable NIC". A caller that believed otherwise would
-    /// be reasoning about a network path that does not exist.
-    #[test]
-    fn libkrun_capabilities_declare_the_vsock_only_shape() {
-        let caps = LibkrunBackend.capabilities();
-
-        // The NIC-less posture, stated both ways.
-        assert!(!caps.tap_networking, "a libkrun workload guest has no TAP");
-        assert!(
-            caps.no_routable_guest_nic,
-            "a libkrun workload guest has no net device to route at all"
-        );
-        assert!(caps.vsock, "vsock is the only transport into the guest");
-        assert!(caps.host_vsock_proxy);
-        assert!(
-            !caps.l3_vsock,
-            "libkrun does not meet the stricter L3 tunnel precondition"
-        );
-
-        // The rest of the table, each pinned so its deletion is visible.
-        assert!(
-            !caps.pause_resume,
-            "libkrun's C API exposes no pause/resume"
-        );
-        assert!(!caps.snapshots, "libkrun has no memory snapshots");
-        assert_eq!(caps.snapshot_capability, SnapshotCapability::DiskOnly);
-        assert!(caps.standby_pool);
-        assert!(!caps.balloon, "libkrun's C API exposes no balloon control");
-        assert!(
-            !caps.fs_quick_checkpoint,
-            "the libkrun rootfs is a regular file, not an APFS clone source"
-        );
-    }
-
-    /// A backend's liveness probe decides whether a VM reports Running or
-    /// Stopped. Pinned to `true` a dead VM is Running forever, and the
-    /// reaper never collects it; pinned to `false` a live VM looks gone.
-    #[test]
-    fn pid_alive_distinguishes_a_live_process_from_a_dead_one() {
-        // This process is alive by construction.
-        assert!(pid_alive(std::process::id() as libc::pid_t));
-
-        // A child we reap ourselves is definitively gone afterwards.
-        let mut child = std::process::Command::new("/bin/sh")
-            .args(["-c", "exit 0"])
-            .spawn()
-            .expect("spawn a short-lived child");
-        let pid = child.id() as libc::pid_t;
-        child.wait().expect("reap the child");
-        assert!(!pid_alive(pid), "a reaped child must not report as alive");
-    }
-
-    /// The timestamp stamped onto libkrun VM state. A constant makes every
-    /// VM look like it started at the same moment, which defeats any
-    /// age-based reaping built on it.
-    #[test]
-    fn now_unix_secs_is_a_real_clock() {
-        let observed = mvm_core::time::now_unix_secs();
-        // 2020-01-01, comfortably in the past and comfortably after any
-        // constant a mutation would substitute.
-        assert!(
-            observed > 1_577_836_800,
-            "expected a current unix timestamp, got {observed}"
-        );
-        assert!(
-            observed < 4_102_444_800,
-            "expected a timestamp before 2100, got {observed}"
-        );
-    }
-}
