+++ crates/mvm-backends/src/lib.rs
@@ -6,6 +6,5 @@
 
 pub mod driver;
 pub mod fc;
-pub mod legacy;
 
 pub mod mock;
